Disable source jar navigation by default.
For now, just disable it under an experiment. If this doesn't cause problems, I'll remove it entirely in a later release.

It's no longer necessary, and is actively making things worse. Instead of redirecting via ClsCustomNavigationPolicy, rely on BlazeAttachSourceProvider, which will actually attach the source jars to the library, so they're properly indexed by the IDE.

PiperOrigin-RevId: 249071475

GitOrigin-RevId: 2e51cc0c86f8488c4110ba616532516e3b9d4a91