Improve performance retrieving experiment data
- keep a global cache of hashed keys
- never calculate experiments list synchronously, instead refresh list every 5 minutes.

This means users will need to start an experiment scope (e.g. by syncing), restart the IDE, or wait 5 minutes to see experiment changes, but that's an acceptable cost.

If we eventually want to improve this, we could have the ExperimentLoaders trigger a refresh in ExperimentServiceImpl when they detect changes (e.g. a ~/.intellij-experiments VFS change).

PiperOrigin-RevId: 249844245

GitOrigin-RevId: 5a072a54f139ba9aaf111319619e980026aab64d