Ensure target is up to date in android run config handler.
Due to the way blaze run configuration editors update the underlying
configuration state, it's possible for the configuration referenced in
android run config handlers to be out of date. This can cause tricky
side-effects such as incorrect build target being build during test.

This CL ensures the target is up to date when a new runner is created
from the handler by updating it from the execution environment.

Validation of the updated target label is not needed here because:
1. The target kind is guaranteed to be a support target kind of the
   handler or else it wouldn't have been used in the first place.
2. Any other validation is done during edit-time of the run
   configuration before saving.

PiperOrigin-RevId: 254436613

GitOrigin-RevId: add2ec405e2da9b3858f75f479206d726fc96472