Cleanup: package target data together as a top-level BlazeProjectData message.
The only reason to put it in SyncState was to hide the aspect-specific implementation details, back when there was a possibility of implementing the IDE-blaze interface differently.

We've well and truly settled on using an aspect, and the ProjectTargetData state already exposes BlazeIdeInterfaceState outside the 'aspect' package, so time to move it out of SyncState.

Future work: if we ever open-source the repackaging step, I'd like to make SyncState truly extensible in the proto, replacing that message with a 'repeated google.protobuf.Any' field so that sync plugins can add whatever they want to the proto.

PiperOrigin-RevId: 257240001

GitOrigin-RevId: 257f9a0f3287d2012f1f90305cbfcd9652c7ac99