Expose manifest_values field in AndroidIdeInfo proto wrapper
Users can override attributes of the merged manifest in their BUILD files using the manifest_values argument of the android_binary and android_local_test rules. We already track this argument in the AndroidIdeInfo proto.

This CL exposes the manifest_values map in the AndroidIdeInfo proto wrapper so that we can account for these overrides during ASwB's merged manifest computation.

PiperOrigin-RevId: 257821220

GitOrigin-RevId: 4d76eae19eaceaec617ab366239292ff455ea226