Focus Blaze console editor when tool window is focused.
This should already happen by default, but there's a bug in
RunnerLayoutUi's RunnerContentUi. It wraps our content in a GridImpl,
but doesn't override TabInfo's preferred focus component to be either
the root content, or the content's preferred focus component.
So GridImpl gets to decide what should be focused, and always picks
the first empty placeholder cell...

Override all this at a higher level, by making the console the default
focus component of RunnerLayoutUi itself. This is safe, since the only
other component there is the toolbar, which can't be focused anyhow.

PiperOrigin-RevId: 260532582

GitOrigin-RevId: 503aedbd6e3b334a7cba09713938c3c8c0786576