Set the platform prefix based on the plugin_api
Android Studio 3.6+ requires the platform_prefix to be
set to AndroidStudio in order for the bundled Kotlin plugin
to work properly.

Rather than requiring each target to explicitly specify a
platform_prefix based on their knowledge of which platforms
they are applicable to, this CL automatically derives the
prefix based on the current target platform.

PiperOrigin-RevId: 263619224

GitOrigin-RevId: 09e9af2ddc9f6ec711c997eb75408c36e02d532a