Add DEPLOYS_TO_LOCAL_DEVICE user data key to deployable run configs.
The presence of this key in the blaze command run configuration's
user data is required for the device selection combobox to enable
itself when a deployable run config is selected.  Currently the
device selection box never gets enabled for any blaze run config,
even if they are deployable.

Tests: android targets now cause the device selection combo-box to
enable itself.
PiperOrigin-RevId: 269853458

GitOrigin-RevId: 1b461c68cd5978e8d2c23b6b9d836ab304357f24