Guard against NPE when validating instrumentation binary.
Validating a android test run config in ASwB involves obtaining the
instrumentation binary of the test through project target map.  It's
possible for a run config to have stale information and cause it to
NPE when a matching target can't be found.

This CL guards against this with a explicit null check.

PiperOrigin-RevId: 272464096