Add more flexible BUILD file target name matching.
kt_jvm_library currently produces java IDE info with target names
plus a "_DO_NOT_DEPEND_JVM" at the end. This breaks BUILD file target
name matching.

This CL fixes issues like these by augmenting BuildFileUtils#findBuildTarget
to look for a macro with matching prefix when the target isn't in the BUILD
file.

When a label is calculated, #findBuildTarget first checks to see if the
label actually points to a target in the corresponding BUILD file.
If it doesn't, then it iterates through the named macros in the BUILD
file until it finds one with name matching the beginning of the label.

The new macro prefix matching mechanism is enabled by the flag
"buildfile.macro.prefix.matching.enabled"

If all else fails, #findBuildTarget returns null.

PiperOrigin-RevId: 274047598

GitOrigin-RevId: 881250f1118ebdffdaeac720bc515d4251f08114