Fix dlv position converter.
We shouldn't have used a BiMap because the two directions are
different.

For remote to local, we want to keep a normalized remote path to local
file map, so we can convert different remote paths that normalize
identically to the same local file.

For local to remote, we might *not* get the normalized path as a path
recognized by dlv. In that case, we want to return an arbitrary path
that dlv recognizes and that we have mapped to the input local file.

Only when dlv recognizes the normalized path itself is the mapping
bidirectional. Otherwise the two maps are not the same.

PiperOrigin-RevId: 279754760

GitOrigin-RevId: 2ef294b76b1113c463a24d3af98bdfd6dcce9619