Do not inject services via the constructor
One restriction with sdk v193 is that services should not be injected
via the constructor. There is also a new annotation (@NonInjectable)
that needs to be placed on constructors that are not supposed to be
used by picocontainers.

This CL adapts the fast run code to work under these restrictions.

PiperOrigin-RevId: 280033808