Add tests for blaze APK build steps and BlazeApkDeployInfoProtoHelper.
There are currently no test coverage for blaze APK build steps. The
build steps are a critical part of run/deploy pipeline and regressions
in this area are guaranteed to be severe.

This CL adds tests tha cover the existing build steps.  This CL also
includes updates to BlazeApkDeployInfoProtoHelper to make it more
testable, as well as a unit test for the proto helper itself.

Expanded visibility on some blaze-intellij classes because they were
required to write these tests.

PiperOrigin-RevId: 284639809

GitOrigin-RevId: 5e0de4dc9936f744abc9da9a59df0bf83c04c371