Use LinkedHashMap instead of HashMap to store libraries.
Classpath in IntelliJ is determined by the order in which JARs are added to the
module. We store the libraries in a HashTable and then stream it into a List,
which effectively radomizes the order in which JARs are added. This causes
potential symbol resolution issues where a dependent JAR packs a subset of its
dependecies, and IntelliJ uses the dependent JAR as canonical source of the
dependency.

This CL changes the HashMap to a LinkedHashMap which preserves the order in
which keys were added. As a result, streaming the JARs into a List retains the
order in which JARs were added to the HashMap. JARs are added to the
LinkedHashMap in a bottom first order. The dependencies are added first
followed by the dependent JARs. This prevents any conflicts due to improper
packaging of the dependent JARs.

PiperOrigin-RevId: 284996633

GitOrigin-RevId: f2e1cb5ddb1272b391eebf0d089768c2482fef63