Repackaging uses constant timestamps for files to make jars reproducible
The CL updates the process of repackaging jar files with a new step that
sets all the creation, modification and access dates to a constant for
all the files inside of an output jar. This should allow multiple
repackaging of the same input jar file to produce exactly the same
outputs.

PiperOrigin-RevId: 288500617

GitOrigin-RevId: 752069048daae61db33a2636d87a822e3a473ef8