api193: Java integration test fixes.
api193 expects `org.junit.Test` and `org.junit.runners.JUnit4` classes to be present in the project to classify it as a JUnit4 project. CombinedTestHeuristic.java already added the relevant classes, but missed the `.java` extension which caused the tests to fail in 193. This CL adds the mentioned classes to the tests that need it.

PiperOrigin-RevId: 289664095

GitOrigin-RevId: d7c6150fa64e5295f4c6e895e4c12763db677c26