Move where we reload tsconfig files.
After https://github.com/bazelbuild/intellij/commit/9b2bc6f748dc834e2c36e5f91bb3e85d82ca808b, we always reload the tsconfig whenever we read the
file list to avoid having stale configs, but that causes file resolving
to happen on the EDT when we want to rebuild the external libraries for
TypeScript.

This CL moves tsconfig reloading to immediately follow the place where
we rebuild the tsconfig rules. It is off the EDT, and guarantees that
the TypeScript service always has an up to date view of the tsconfig
files.

PiperOrigin-RevId: 291968922