Pretend the tsconfig doesn't have a files list.
According to the specs, if there's not an explicit files list every ts
file under the tsconfig is implicitly included.
This solves the issue where newly created/moved ts files are not
recognized by the tsconfig for path resolution.

We should still keep the actual list for prefetching and resolving
library file symlinks.

PiperOrigin-RevId: 298388373