Retain generated resources from common library dependencies
We automatically exclude generated resources for performance reasons
unless they're explicitly whitelisted in the user's project view. This
shouldn't be a requirement for well-known libraries that rely on
generated resources.

This CL provides a GeneratedResourceRetentionFilter ExtensionPoint
which allows extensions to provide custom filters for retaining
resources from such libraries. WhitelistFilter has been updated to check
the retention filters before checking the whitelist. One consequence of
this is that users will be warned not to whitelist folders that are
already retained by one of the filters (as this is unnecessary).

PiperOrigin-RevId: 302773280

GitOrigin-RevId: 67fef89b129389dee6231f09acf6d0f8e79bab35