Merge resources instead when there are multiple resources modules with the same package.
Currently, when we have multiple targets declaring resources in the same package, we use a heuristic to choose the "best" package. It is not always the best choice and often leads to unresolved symbols in java sources. This CL merges the dependencies and sources of all targets and creates a module with merged resources.

This creates a scope leak, but it should mitigate performance issues due to unresolved symbols (AutoImports, for example).

PiperOrigin-RevId: 304718524

GitOrigin-RevId: 7dc8ff650913421bb53131858fb3e2de4b1e3bc2