Automated rollback of commit c0ba4dc3b8a320cdd8585d4c80731774f5074d68.
*** Reason for rollback ***

Can use ApplicationInfo.getBuild instead

*** Original change description ***

Use sdkcompat to determine if DisableDockerPlugin should be deactivated

The code relied on having access to PluginManagerCore.BUILD_NUMBER, which has
been removed. Rather than exposing the BUILD_NUMBER via reflection, this CL
just exposes a #shouldDisableDockerPlugin that is customized per API.

We may still have to somehow expose the BUILD_NUMBER functionality as some
integration tests may rely on it, but that'll be fixed when we come to that
failure.

PiperOrigin-RevId: 314621478

GitOrigin-RevId: 6b296ab576c00dcf370180489104bc654c1d2990