Automated rollback of commit 2d13d1f0464b6c3a3bb11468007733301aea4f40.
*** Reason for rollback ***

Can use ApplicationInfo.getBuild instead

*** Original change description ***

Use sdkcompat to determine if DisableDockerPlugin should be deactivated

The code relied on having access to PluginManagerCore.BUILD_NUMBER, which has
been removed. Rather than exposing the BUILD_NUMBER via reflection, this CL
just exposes a #shouldDisableDockerPlugin that is customized per API.

We may still have to somehow expose the BUILD_NUMBER functionality as some
integration tests may rely on it, but that'll be fixed when we come to that
failure.

PiperOrigin-RevId: 314621478