Add e2e test for android deployment and launch tasks.
ASwB's run/debug area is missing tests that cover the overall
flow from creation of a run configuration to the generation of
APK deployment and launch tasks. There exist tests that verify
the individual steps through out this flow but breakages in the
overall integration can't be caught by these existing tests.

This CL adds an end-to-end test with coverage from the creation of
a run configuration to the generation of the APK deployment
and launch tasks.  Note this CL only adds one test and it
only covers running an android_binary target with mobile-install
disabled.

Following this change we should add tests that cover permutations of:
* Executing with different executors such as Debug and Profile.
* Building with other launch methods such as mobile_install.
* Handling other kinds of targets such as android_instrumentation_test

PiperOrigin-RevId: 328671373

GitOrigin-RevId: 990bd690794084fbdaaedaf564f0e2f4fb94cc38