AarLibrary: attach source jars for aars
aar_import rule now allows specifying a source jar. This CL attaches
the source jar to the IntelliJ library corresponding to the aar.

Currently, we always attach these sources, which differs from the
java plugin's behavior of only attaching sources when navigating
into those jar files. Doing so for aars requires a little bit of
restructuring as all the entry points into attaching sources at the
right time assume that the sources are part of a BlazeJarLibrary,
rather than an AarLibrary.

PiperOrigin-RevId: 328853483