Passing --noexperimental_run_validations to avoid unnecessary valiation actions
We don't need validation actions to run at all for sync. Earlier we had no way to turn it off, but the new flag provides a way. Even though most of the time these actions are not on the critical path, every now and then they are. So turning it off is worthwhile for sync latency.

With this cl, --noexperimental_run_validations will be set when sync projects
to avoid validation action to run. This flag will not impact analysis cache, i.e. it should not affect build performance.

PiperOrigin-RevId: 332890574

GitOrigin-RevId: d6c1e9bdbf0b8ab225d387ffe51f8341b6b25bbc