Add logs in UpackedAars.getClassJar when fails to find remote artifact from cache state
When BlazeArtifact is remote one, we would expect it can be looked up from cache state. If not, a run time exception will be thrown. However there's no artifact information is logged when the runtime exception is thrown. Then it's hard to debug it if we cannot reproduce it locally.

With this cl, a log is added before thrown exception. It includes library artifact information and cache state keys.

PiperOrigin-RevId: 333568420

GitOrigin-RevId: 581dec7852f6acaef4249bf74af7c3d0aba8546e