Improve post-sync errors and workspace module creation
Currently there are many ways for BlazeAndroidProjectStructureSyncer
#updateProjectStructure to silently fail. In some of these silent
failures the workspace module won't be given an AndroidFacet which
will cause a variety of errors down the line.

This CL adds user facing error messages that informs the user about
these errors and how to solve them. Workspace module creation code
has also been moved earlier so the AndroidFacet is always initialized
as long as it's an android project regardless if the sync failed.

PiperOrigin-RevId: 333643559

GitOrigin-RevId: c2d5159ae9b6f0a4dc38327f47d67f1108c7ba76