Attach both java and native debuggers when native debugging is enabled
Currently when native debugging is enabled in a run configuration and
the debug action is triggered on that run configuration, the app will
launch with only the native debugger attached.  This CL changes usage
of BlazeNativeAndroidDebugger to BlazeAutoAndroidDebugger so that both
java and native debuggers will be attached.

Additional workspace path and source-map configuration code has been
moved from BlazeNativeAndroidDebugger to BlazeAndroidDebuggerService
because both BlazeNativeAndroidDebugger and BlazeAutoAndroidDebugger
uses it.

PiperOrigin-RevId: 334842418