Change blaze native debugger ID from "BlazeNative" to "NativeBlaze"
Due to an upstream bug in Android Studio, debugger selection will
always default to the first debugger in a list of available debuggers
sorted by their debugger ID.  This means the "BlazeNative" debugger will
always be selected instead of the "Java" debugger. We should default to
the java debugger because it's much more commonly used than the native
debugger.

This CL changes BlazeNativeAndroidDebugger's ID from "BlazeNative" to
"NativeBlaze" so it comes after the "Java" debugger. Before the bug is
fixed upstream, "Java" debugger will be the default because it's the
first debugger.

Note that the debugger's ID is used by AndroidProcessChooserDialog to
remember which debugger was used when no run configuration is given.
This will be updated upon first selection of a debugger after this
change and has no other impact on debugger selection or run configs.

PiperOrigin-RevId: 335131634

GitOrigin-RevId: ebdeef97419f46f6cdd811427c2f6e5d4a9fbc99