Allow model properties to be observed through listeners
Architectural patterns like Model-View-Controller, Model-View-Presenter, and Model-View-ViewModel require propagation of changes of model properties to the UI implementation. Of course, the code updating the model (e.g. the presenter) could afterwards also update the view with the same changes. That's tedious and error-prone, though. The typical approach (in a Java Swing or JavaFX world) is to use listeners which attach to model properties and update the corresponding UI parts when triggered.

By itself, Java Swing offers the PropertyChangeListener class and allows to attach PropertyChangeListeners to various UI classes like JLabel to observe when specific properties of them change (e.g. the text of a JLabel). The issue with PropertyChangeListener is that this class is a very old implementation (from Java 1.1). It doesn't benefit from any newer developments in the Java world (e.g. no generics). In addition, its design is based on older approaches which are not up-to-standard anymore. For instance, a PropertyChangeListener can't be attached to a property directly. Instead, the listener attaches to the UI class (e.g. the JLabel) and needs to filter on the property name which is passed as string. Furthermore, PropertyChangeListener makes it difficult to use newer, recommended approaches like data bindings (which build on such listeners) or conversion of an observed value through a passed method reference.

As Java Swing is pretty old, hardly used anymore, and wasn't developed further within the last years, there are no handy, shiny, maintained libraries for such a listener/binding framework which we could simply use.

JavaFX which was meant as successor to Swing brought its own framework of listeners and data bindings. That framework was largely improved compared to the one Swing offered out of the box. Considering that JavaFX is available as a Java library by now, we could theoretically use that framework. However, it's optimized for use within JavaFX (e.g. automatic check of the use of the JavaFX application thread vs. the EDT of Swing) and it offers many parts (e.g. lazy evaluation depending on whether a UI property is bound) which we simply can't properly use in a Swing application. On the other hand, the general idea of JavaFX's listener/binding framework would be helpful for our purposes as well. Therefore, we introduce a simple variant of such a listener/binding framework which is inspired by JavaFX. This CL only introduces the listener part. The binding part will be part of a follow-up CL.

The basic idea is that all model properties are fields of type Property. Every Property allows to register listeners and notifies them when its value changes. There are two types of listeners: InvalidationListeners and ChangeListeners. InvalidationListeners are triggered on every change. ChangeListeners are only triggered when the new value is different, which means an additional equals() check.

Bindings are typically based on InvalidationListeners as they just care about value propagation. In addition, bindings may have intermediate conversion steps because of which an originally unequal value may be equal after the conversion, rendering the use of a ChangeListener meaningless. That's why we add both types of listeners.

As JavaFX, we also add the separation between an observable value (-> read focused) and a modifiable property (additionally offers write functionality). As long as there's no need to offer special implementations of a property, we directly implement this class.

PiperOrigin-RevId: 347600958

GitOrigin-RevId: 28b65dc76a9814c26c4721ca6ac3328da3f9576b