Use image diffing in AswbRenderTaskTest
Currently AswbRenderTaskTest uses view component properties to verify if
a layout was rendered correctly.  Some components of a rendered layout
cannot be verified through view properties.  E.g. Correct rendering of
an image or icon.

This CL adds image diffing to AswbRenderTaskTestl mostly as an example
for future tests that will need image diffing for render tests for the
blaze android plugin.

AswbImageDiffUtil is added and provides image diffing utilites that can
be used in these tests.  AswbImageDiffUtil is a copy of ImageDiffUtil.kt
before it was converted to kotlin.  Here we use the java version because
kotlin support is quite complex at the moment and outside the scope of
this CL.  AswbImageDiffUtil should be updated to the kotlin version when
appropriate.

PiperOrigin-RevId: 348539922

GitOrigin-RevId: 25b093ec70aaf038ee16f9126a348e0378bda52c