Check both internal and external SDK paths when initializing layoutlib
Currently AswbRenderTestUtils#getStudioSdkPath only checks the internal
repository path for the SDK and copies the required files if found. This
fails when the test is run from external github repository because the
SDK location is different.

This CL fixes this by checking both internal and external paths when
initializing the SDK to ensure the required files can be found in both
scenarios.

PiperOrigin-RevId: 351210891