Remove assumption that every manifest must define a package name
Currently the ManifestParser assumes every manifest file is required to
have a defined package name and is otherwise a malformed manifest. This
is only true for final merged manifests. There can exist manifest in
the user's sources that only define some aspects to be merged into the
final manifest; a package name is not required for these manifests.

PiperOrigin-RevId: 355436989

GitOrigin-RevId: f0702730245e60a94b10367ee3e911ee5a888e3b