BlazePackageSearchScope: Override correct method for hashcode implementation
SearchScope.hashcode() is marked as being non-extendable. According to the
Javadoc description, calcHashCode() should instead be overridden to provide
a custom hashcode implementation whose result can be cache. Caching seems
to be necessary for performance reasons especially for Kotlin projects
(see commit message of the change in the IntelliJ platform:
https://github.com/JetBrains/intellij-community/commit/ecc334ddae0ecaee401cb906ea32089e1a92ee09).

This issue was flagged by the IntelliJ Plugin Verifier.

PiperOrigin-RevId: 357726823

GitOrigin-RevId: 5d660ed669408a7e136c3c1e9815d32e9f9f707a