Fix mistake in ExternalDependencyMigratorTest
during the creation of the test, it was not considered that `file.getRoot()` will always return Path yet `ConversionContext.getSettingsBaseDir()` return type will depend on the Sdk version. To be precise, it will return Path in 2020.3+ and File before. This causes the test to fail on 2020.3.

To fix this, we introduce a slim wrapper method that will transform file.getRoot() to Path in 2020.3 and not do anything in older versions. The method can be cleaned up completely once we pave 2020.2.

PiperOrigin-RevId: 362776225

GitOrigin-RevId: 1224a11c1cee93a27f831705fe6f5517e6b74715