Map /proc/self/cwd to workspace-root for all OSes on native debugging
Currently BlazeNativeAndroidDebugger#createState remaps /proc/self/cwd
to the workspace-root on MacOS because auto resolution of /proc/self/cwd
only works on linux. A recent update to LLDB removed the setting that
allows the automatic resolution to work on linux, so now this custom
remapping is required for all OSes.

PiperOrigin-RevId: 364690259

GitOrigin-RevId: 989c6e6ba7089de4b50f5be3ee07abcc10b487e4