Fix unexpected behaviour when running tests from a directory.
When running tests by clicking on the directory with subdirectories, the default assumption should be all tests inside that directory + all subdirectories should be executed. However, the reality is those within the subdirectories are skipped. This unexpected behaviour caused us problems when we didn't have a TAP project and our submitted code broke some code without us knowing (we did test it locally but nested tests were skipped).

This CL changes this behaviour to running all tests in all subdirectories when clicking into the root directory, and skipping tests from subdirectories when clicking into the BUILD file, which is compatible with the behaviour of Partial Sync.

Limitation: After some search, the directory to click must also be a Blaze directory (with a BUILD file) because of interference with the MultipleJavaClassesTestContextProvider (a Java test provider) which somehow runs after the Blaze test provider (it should have been matched before). This is, however, still useful since most folders in Google3 have a BUILD file anyway and especially for non-Java languages (like Go) where code and tests are in the same place.
PiperOrigin-RevId: 367498549

GitOrigin-RevId: bdcc72926835e3470c0882c35667f0c8d997b567