Suppress warnings for overridden methods lacking generic types due to SDK compat
Suppressing the 'unchecked' warnings directly on the methods doesn't
help for the 'unchecked' warnings about the method overrides. Hence,
also suppress the warnings on a higher level. We leave the suppression
on the methods also in place to better highlight the methods causing
the warnings.

In BlazeGoRunConfigurationRunner, we extract an inner class so that
the scope of the warnings suppression is as limited as possible.

PiperOrigin-RevId: 367653720

GitOrigin-RevId: 530086ab4429a45906425af3885d7c8e33d90a2e