Notify that roots changed after external library manager initialization.
In a normal blaze sync we initialize external libraries in updateProjectStructure, which would also modify project roots and libraries and trigger a roots changed event and force the IDE to reindex files.

We don't run updateProjectStructure in the startup sync since we don't expect source roots and libraries to change from reopening the project, so we had special handling for the startup sync in ExternalLibraryManager using a SyncListener.

2020.1 and prior it was sufficient to just initialize the additional library roots provider there, but after 2020.2 we also need to notify the IDE that roots have changed.

PiperOrigin-RevId: 368447211

GitOrigin-RevId: 0a40bf6cde1570eb48f882ace304bcdae79bec85