Return canonical resource module for targets of merged resource modules
Currently AndroidResourceModuleRegistry keeps a one-to-one relationship
between a resource generating target target, it's AndroidResourceModule,
and its corresponding intellij Module. This model cannot accurately
represent resource dependency relationships when a target's dependencies
generate resources under the same resource package name.

https://github.com/bazelbuild/intellij/commit/71dc54bb49ce633f5b5fae2c7fe5e84bc4a9f624 introduced resource module merging which provided a way
for targets of the same resource package name to contribute resources to
one single AndroidResourceModule.  AndroidResourceModuleRegistry keeps
track of this mapping but it was not updated to account for the merging
of these AndroidResourceModules.  This lead to errors with finding a
resource's dependencies.  E.g. a dependency's label will not correspond
to an AndroidResourceModule, but its resources were absorbed during
merging; so there is a resource module, just that it's identified with a
different label.

This CL fixes AndroidResourceModuleRegistry and surrounding resource
module creation logic to correctly interpret these resource module
dependencies.

PiperOrigin-RevId: 368495669

GitOrigin-RevId: 015c1a86ca15dcd3473ce03b2b9113225fe8bd07