Replace JavaInfo.transitive_exports with data collected by the IDE aspect (IntelliJ).
Collection of the data in the aspect directly is better, because it handles not only Java targets, but also Java-like targets (like kotlin).

We are also looking to remove JavaInfo.transitive_exports and JavaExportsProvider from Bazel, because collecting labels in a provider is unusual and the implementation is incomplete for Java-like targets and it's impossible to make it complete without awkward extensions to the APIs.

Design doc: https://docs.google.com/document/d/10isTEK5W9iCPp4BIyGBrLY5iti3Waaam6EeGVSjq3r8/edit#bookmark=id.34gjanvu5611

PiperOrigin-RevId: 369661282

GitOrigin-RevId: 6762db98367e470a58c9765e5bdc2005c32c57a4