Allow to run non-small Kotlin tests defined in general test rules
We have various heuristics kicking in when determining which target
to choose for a given file. If the Kotlin file was directly referenced
as source in the build target, users could run tests for it in IntelliJ.
However, if test rules were generated according to size categories
(e.g. SmallTests, MediumTests), IntelliJ always tried to execute the
SmallTests no matter which test size was actually configured for the
tests in the code.

This behavior was based on the fact that we didn't derive the test size
for Kotlin tests yet simply because it was not implemented. If no test
size is given, the heuristic which maps a test to its target according
to the test size assumes that the test is small.

This CL adds the necessary code to derive the test size for Kotlin
tests. It's based on the implementation which identifies the test size
for Java tests. Unfortunately, we needed to re-implement the logic for
Kotlin as the PSI class hierarchy of Kotlin and Java don't overlap.

We're also adding quite some tests covering the new behavior. Similar
tests don't seem to exist for Java. Hence, we're not branching from
them.

PiperOrigin-RevId: 372911532