Invoke BlazeSyncPlugin.updateInMemoryState w/o a read lock
BlazeSyncPlugin.updateInMemoryState is invoked while a read lock is held. This
makes it harder for the implementations of that extension point to run long
running actions, parts of which may not need the read lock.

This CL moves the acquisition of the read locks to be done inside the implementations,
allowing them to further customize when read locks are acquired in the future.

PiperOrigin-RevId: 374278077

GitOrigin-RevId: e6503f6816f4e9a1a5b97d26075dc8728c783e4e