[2021.1 compat] Add Compat functions for TypeScriptConfigService interface
The TypeScriptConfigService interface was updated in 2021.1
* Methods `getConfigs` and `hasConfigs` were removed
* Methods `getPreferableOrParentConfig`, `getDirectIncludePreferableConfig` and `getRootConfigFiles` were added

To support older IDE versions, we provide implementation for all these methods in classes implementing TypeScriptConfigService while omitting the @Override.

The newly added methods are not implemented yet but we need this CL to move past build errors caused by it, we have b/188883814 to track the API support.

PiperOrigin-RevId: 375655045

GitOrigin-RevId: 6da758a34859c93a49be731ac4193ef3583ece2b