Remove instance of BlazeCommandRunConfiguration in android run handler
Currently BlazeAndroidBinaryRunConfigurationHandler and
BlazeAndroidTestRunConfigurationHandler stores an instance of the
BlazeCommandRunConfiguration the handler was created to handle. The
BlazeCommandRunConfiguration maybe changed during run configuration
editing, making the stored instance point to the incorrect run
configuration. The correct way to obtain a BlazeCommandRunConfiguration
is to extract it from the execution environment.  This CL fixes this.

PiperOrigin-RevId: 376027546

GitOrigin-RevId: 8bd332f3a55f81b71f590bd86edc589914aa1ba3