Extract deployement info from deploy task instead of internal context
A part of AndroidBinaryNormalRunTest verifies that the correct APKs
are selected for deployment.  Currently this involves using
BlazeAndroidBinaryNormalBuildRunContext#getApkInfo which was exposed
for the sole purpose of testing.  AbstractDeployTask offers a public
API that achieves the same thing.  This CL switches to that public
API instead of a test-only internal API.

PiperOrigin-RevId: 377098895