Make application id computation async
We originally computed the application id on-demand. This resulted
in freezes in some projects, specifically when the "attach debugger"
action was invoked. https://github.com/bazelbuild/intellij/commit/bc60b0833f82b35f01b06f3f9904d09f918cc416 was the attempt to fix this by
always computing the application id for all modules at sync time.

This introduced a major freeze at sync time - now we are needlessly
computing the application id for all modules (including hundreds of
resource modules), and reading the manifest out of srcfs can be
slow, esp. on laptops where there is no persistent srcfs cache.

https://github.com/bazelbuild/intellij/commit/993ae7ab552d9ae49ddc062e478e8229b25107c0 was an attempt to confirm that the majority of the time
went into manifest parsing, which was indeed confirmed to be the case.
This CL removes this addition as we don't need it anymore.

This CL just pushes the manifest computation to a separate task
in an executor. This means that there should be no freezes at sync
time, although it is possible that some actions that ask for the
application id can see a freeze if the id hasn't been calculated yet.
This is currently limited to a max of 1 second freeze though.
Callers will instead see a special application id for
"application id unavailable", which is supported in AS 2020.3+.

PiperOrigin-RevId: 380279805

GitOrigin-RevId: d6789244a60d855cc740ed41a659fb9979ea8c17