Make application id computation async
We originally computed the application id on-demand. This resulted
in freezes in some projects, specifically when the "attach debugger"
action was invoked. https://github.com/bazelbuild/intellij/commit/72ae1a6604ef6523b0ff2b0a94df5e99699b4540 was the attempt to fix this by
always computing the application id for all modules at sync time.

This introduced a major freeze at sync time - now we are needlessly
computing the application id for all modules (including hundreds of
resource modules), and reading the manifest out of srcfs can be
slow, esp. on laptops where there is no persistent srcfs cache.

https://github.com/bazelbuild/intellij/commit/ef316e7718dee3ceeda9f4d7e71fff24e2b6d422 was an attempt to confirm that the majority of the time
went into manifest parsing, which was indeed confirmed to be the case.
This CL removes this addition as we don't need it anymore.

This CL just pushes the manifest computation to a separate task
in an executor. This means that there should be no freezes at sync
time, although it is possible that some actions that ask for the
application id can see a freeze if the id hasn't been calculated yet.
This is currently limited to a max of 1 second freeze though.
Callers will instead see a special application id for
"application id unavailable", which is supported in AS 2020.3+.

PiperOrigin-RevId: 380279805