Special case FileNotFoundException
Many of our logs are littered with an exception that arises in the
following scenario:

 - user does a sync
 - close project
 - some time later, do a sync and then open the project

At this time, we do a startup sync, which relies on old ide info
with pointers to manifest files. Some of these manifest files may
not exist anymore in which case we log a large exception in the
log file. This CL just changes it to be a smaller warning in the
log instead.

PiperOrigin-RevId: 380281873