Move external library initializations off the EDT.
We previously assumed that resolveVirtualFile was okay on the EDT if we didn't refresh files, but it still does a bunch of getAttributes calls which can slow things down.

We can initialize the library files in refreshVirtualFileSystem (off EDT) instead of in updateProjectStructure (on EDT). Since updateProjectStructure comes right after refreshVirtualFileSystem, indexing is still triggered on the external library.

PiperOrigin-RevId: 381864823

GitOrigin-RevId: e7c3fef23669fa778d41a7c6925b368a215882fd