Use new target map builders in BlazeAndroidWorkspaceImporterTest
BlazeAndroidWorkspaceImporterTest uses target IDE info builders to build
the target map which is very verbose and hard to read.  This CL changes
it out to the target map builders made for tests that are much easier to
read.

NbTargetMapBuilder and related classes are moved out of integration
testing utils because it's used in more than just integration tests.

PiperOrigin-RevId: 382564561

GitOrigin-RevId: 9b9a76961db8660c11451bb3021cfac744ac468d