Update stamp plugin xml to handle EAP versions
Not all plugin api versions have lib/resources.jar from which we used to get the application information to set the idea-version compatibility range of the plugin.
However, `product-info.json` file contains this information but it was not imported to google3 with previous plugin api versions.

This CL introduces `application_info_container` file group which contains the `application_info_jar` for previous versions and `application_info_json` for new ones. It also updates the xml stamping script to use `application_info_container`.

Additionally, this CL updates the way the range of compatible IDE versions is generated for EAPs. For example with build number: IU211.6693.111
- If it is a stable version: since=211.6693, until=211.* (same as it is now)
- If it is an EAP version: since=211.6693 until=211.6693.* (to make sure that this version is not available for the official release of 211 when it is out because it may contain other breaking changes).

PiperOrigin-RevId: 382693979

GitOrigin-RevId: 26f031140f0ba98e22fb61129c2b49644d39f363