Fix NPE in BlazeXmlToTestEventsConverter#parseComparisonData
The message can be null as e.g. null checks in
BlazeXmlToTestEventsConverter#pruneErrorMessage show.

We additionally checked other usages of ErrorOrFailureOrSkipped.message
and added an additional null check for safety in
BlazeXmlToTestEventsConverter#processTestCase. This might not have been
necessary and is unrelated to the reported NPE but could have
represented another potential NPE.



PiperOrigin-RevId: 389914193

GitOrigin-RevId: 0c253d2d554f9b33e68ed3afacbd19ee716ea85e