Fix BuildEventServiceUploader to always tolerate errors on stream half-close.
Previously if an error is produced by the server after half-close, and all acks
are already received, the uploader will re-open the stream but then have no more
work to do: all events were sent and acked. The client will then hang with the
open stream.

This change allows for this possibility and bypasses client retries, bubbling up
the error returned by the server.

.

PiperOrigin-RevId: 389987070

GitOrigin-RevId: c2b2e2162ca145acd7c95d2df4b6684334728a41