Add options to stamp_plugin_xml to specify the supported build numbers range.
This CL adds two options to stamp_plugin_xml; since_build_numbers and until_build_numbers. Each takes a map from a major IDE version to the value of since build number or until build number to be used in the plugin xml file respectively.

Each of the two options can be used independently of the other. If no value is specified for a certain IDE version, the default will be used:
- For since build number, it will be the major version; e.g. 212
- For until build number, it will be all minor versions of a major version; e.g. 212.*

PiperOrigin-RevId: 395690928

GitOrigin-RevId: 41f60ce4c213784578f542fc1158e23ca0f19cdf