Make sync tests independent of the default Java language level
Switching the default Java language level should not break unrelated tests. As Java-based sync tests need a mock JDK to work, we need to ensure that it's present in the test environment. So far, setting up the JDK was hard-coded by always providing a JDK 8 (corresponding to the current default Java language level) in BlazeIntegrationTestCase.

PiperOrigin-RevId: 395879739

GitOrigin-RevId: 20b0fd2a544ff5f8b88000c3b4b3d02081b52f0f