Move visibility definition outside BUILD files
This CL creates package groups used in targets visibility outside the BUILD files. We will have two versions of the build-visibility.bzl file one to be used internally and one for the external use. This should simplify the copybara transformations applied on the targets visibility attributes which are not always straight forward to reverse.

PiperOrigin-RevId: 395886422

GitOrigin-RevId: da3f08c35eea3ee5c3a0748fc83cacca5a082b25