Avoid warnings by using lenient stubbing for general setup of LoggedDirectoryProvider tests
In those tests, we deliberately set the build system by default for all tests to provide a reliable test environment for tests which don't care about the build system explicitly. Some of these tests don't end up even calling the line of code which requires that a build system is set. This results in "[MockitoHint] Unused XXX.setBuildSystemTo" warnings. We can avoid them by specifying the stubbing as being lenient.

PiperOrigin-RevId: 395899244