Avoid warnings by using lenient stubbing for general setup of LoggedDirectoryProvider tests
In those tests, we deliberately set the build system by default for all tests to provide a reliable test environment for tests which don't care about the build system explicitly. Some of these tests don't end up even calling the line of code which requires that a build system is set. This results in "[MockitoHint] Unused XXX.setBuildSystemTo" warnings. We can avoid them by specifying the stubbing as being lenient.

PiperOrigin-RevId: 395899244

GitOrigin-RevId: 7cc6d967c4c0d7e7ea798e2d1e8b98bb09460a9a