Add Kotlin library first before adding other libraries
IntelliJ 2020.3, when used with the old project model, raises
many root change events for adding a single library to a module,
which in turn results in a long freeze at the end of a sync
when the Kotlin sync plugin adds the kotlin runtime library
to the workspace module's dependencies.

In order to avoid this, this CL introduces a workaround specific
to the time when the old project model is in use: In such a case,
we add the Kotlin library before committing ModifiableRootModel.
Since the number root change events is proportional to the number
of libraries already added to the module when getting
ModifiableRootModel of it. This workaround reduces the time of KotlinWithLibraryConfigurator.configureModule from 70k ms to less
than 300 ms.

PiperOrigin-RevId: 398549044