Ensure that RJCFF.findClassFile never throws an exception
The contract for the method findClassFile requires that it never
throw an exception. We've failed to meet this contract a couple
of times now, which then leads to users not being able to even
open files. In order to avoid this situation in the future, this
CL wraps the call in a try-catch to ensure that no exceptions are
thrown.

PiperOrigin-RevId: 399800576