ArtifactCache: constrain entries to be OutputArtifacts
The ArtifactCache (used only by the RenderJarCache for now) currently supports
BlazeArtifacts which can be sources or output artifacts. This unnecessarily
complicates the code as there is really no need to cache SourceArtifacts.
This CL simplifies the code by making the cache only support OutputArtifacts,
which are artifacts that are produced by a build.

PiperOrigin-RevId: 401095425

GitOrigin-RevId: c4e8341fe652a677dcfdc02b130a44ca100e2fb5