Use relative resources paths
This CL reorganizes the base package of Bazel plugin to allow using relative paths to refer to resources in code. This is done by moving base/resources folder to base/src/resources.

Manually tested that with the new structure, icons, buildifier binary and color schemes can be retrieved successfully for both Blaze and Bazel plugins.

This CL applies the modification requested in unknown commit which was reverted by mistake.

PiperOrigin-RevId: 404552911

GitOrigin-RevId: 8b1b856b709d7df33ca8d321e0c890bc29cdc2d6