Switch invocation policy enforcer to use behavior enum instead of the old boolean flags.
We have recently added a new enum to express the intended behavior for `SetValue` invocation policy entries. This enum replaces the `overridable` and `append` boolean flags and offers significant advantages over those -- does not have a usable default value, thus must be explicitly set, and does not allow combination of `overridable && append` which is semantically contradictory (`overridable` only applies to non-repetable flags and `append` to repeatable flags).

Deprecate the boolean flags which are superseded by the new enum and use the enum instead of the boolean flags in invocation policy enforcer.

PiperOrigin-RevId: 405924139

GitOrigin-RevId: 528d94737ba36721bf5b6bd6e751a49cf0e3f934