Properly handle missing packages and I/O exceptions when parsing target patterns in no-keep-going mode. Additionally, eagerly terminate evaluation on "inconsistent filesystem exceptions" even in keep-going mode, by analogy with a source file changing mid-build, which terminates the build even in keep-going mode. Note that we still don't terminate an evaluation eagerly on all inconsistent filesystem exceptions, but this is incrementally better than before, IMO (I think the remaining holes are bzl files and reading source artifacts for the first time).
TargetPatternFunction can recover from missing packages/other errors in keep-going mode, and so before this change it very rarely threw an exception. However, in a no-keep-going evaluation (or a keep-going evaluation that failed catastrophically), the failure to throw an exception meant that it missed a chance to transform a lower-level exception into a more intelligible one. To allow TargetPatternFunction to distinguish these cases, we add a new method on SkyFunction.Environment.

PiperOrigin-RevId: 415073167

GitOrigin-RevId: be3db92f3695cf82380734ac40f2bc948454e644