Allow existing target.source-map(s) to coexist (bazelbuild/intellij PR import #2974)
# Checklist

- [ ] I have filed an issue about this change and discussed potential changes with the maintainers.
- [ ] I have received the approval from the maintainers to make this change.
- [x] This is not a stylistic, refactoring, or cleanup change.

Please note that the maintainers will not be reviewing this change until all checkboxes are ticked. See
the [Contributions](https://github.com/bazelbuild/intellij#contributions) section in the README for more
details.

# Description of this change

Currently the target.source-map that are defined in `~/.lldbinit` are being overwritten by the plugin. This change allows for all to coexist.

Closes https://github.com/bazelbuild/intellij/pull/2974

PiperOrigin-RevId: 415310654

GitOrigin-RevId: 1bf8d8be4773e092faa40898dc0efdb0859187a4