Build Tasks Window: Group individual syncs under top-level tasks
This is accomplished by wrapping the top level sync tasks (currently at most two: directory update and the actual sync) within a BlazeContext/Scope attached to a ToolWindow Task.

In order to achieve the auto-magic behavior of the progress indicator where the top-level task is marked finished when all subtasks are finished, the tasks within the top-level scope need to run synchronously. To that end, `submitTask` was refactored to await the Future returned from task submission, reducing the number of ignored errors and suppressed warnings.

PiperOrigin-RevId: 431454155

GitOrigin-RevId: 9152aabd9638c9ccf70dc2abc74b6a5dcb593a97