Remove BlazeBuildParams from BlazeSyncParams in favour of deriving it directly from the Project.
In all cases, the BlazeBuildParams are set on BlazeSyncParams by deriving them from the project using BlazeBuildParams.fromProject(). Instead, we can do this where needed, or retrieve the specific info needed from the build params directly from their source where appropriate. Since the build params are determined only by the properties of the project and user settings, this is pure refactoring change and have no impact on the behaviour.

To keep this CL small, we leave BlazeSyncParams.Builder.setBlazeBuildParams() in place, but make it a no-op. Calls to this method will be removed once this CL is merged.

PiperOrigin-RevId: 431906190

GitOrigin-RevId: c378b37a39ce038f2764fc7c9f710456b7284561