BlazeApkBuildStep: Build android_binary targets on rabbit
This CL implements an alternate flow for building android_binary
targets that is specific to Google3/Blaze. In this iteration, it
only supports building on rabbit and then deploying locally via
Studio. It is only enabled in a very limited set of conditions:
the IDE is running off corp, and mdproxy is running w/o sshfs
enabled.

Other types of build (blaze and rabbit via mdproxy) will be
supported once b/225936997 is fixed.

Tests will be added in a follow up CL.

PiperOrigin-RevId: 436469354