Add SyncBinaryType.NONE, and use that by default in SyncStats.
Recent changes means that the SyncBinaryType is not known statically. As a result, we can't populate it in SyncStats in the case then no build was performed as part of a sync (no build binary was used in this case), for example during startup sync.

Tested: Re-open a project after restarting the IDE.
PiperOrigin-RevId: 436726558

GitOrigin-RevId: 2439b056207a1d35a7b9d177d21483014442dcff