Refactor Task to extend PresentableNodeDescriptor
This is the first part of a change to the new Blaze window tree UI to bring it closer IntelliJ's native build window.

This CL does not make any ui changes; Task is refactored to extend PresentableNodeDescriptor, required override methods are added along with the necessary plumbing to ensure that Tasks are initialized with a Project reference.

PiperOrigin-RevId: 437767746

GitOrigin-RevId: 5855bbae1cb61a8f9bbe88c0e853258dc22bf8ec