Bring Task Tree UI closer to native IntelliJ Build window
Refactors TasksTreeView tree to use the StructureTreeModel/AsyncTreeModel backed by Tasks as PresentationNodeDescriptors.

It essentially decouples the TasksTreeModel from the TreeModel powering the Swing Tree, requiring a small change to calculations of TreePaths from a given Task and how we were handling selection of newly added tasks. Previously, the TasksTreeModel selected a new task upon adding it, but with the AsyncTreeModel, this would result in an attempt to select a task before it had actually been added to the tree view. Since we don't want to automatically select all new tasks anyway, the tree handles selection (still delegating to the TasksTreeModel) when tasks are added.

Modifies UI to
- Display duration strings on the right of the tree window while a task is running
- Display finishing time only on the top level task
- Only select newly added tasks when they are top-level tasks
- Auto-expand the first level of tasks

PiperOrigin-RevId: 439664882

GitOrigin-RevId: e2ddd0a8fd8e2191e17ca4f00908bb237a1828f9