Add shortcut ALT+Y for Blaze syncs
As an IntelliJ user, I often have to trigger (incremental) syncs and want to do that comfortably without always using the mouse to click on the icon.

The shortcut is not used for anything else and ALT+Y is not something which can be triggered accidentally (requires wrist movement) unlike S, C, B, or N. Also Y is easy to remember for sYnc.

PiperOrigin-RevId: 440869060