Set sync_flags.fat_apk_cpu to a single cpu architecture when not already set
Building for multiple architectures is costly and also unnecessary for sync operation. This CL avoids this by explicitly forcing the fat_apk_cpu argument to a single architecture when it is not already set by the user. Sync flags are not used during builds or deployment, so this shouldn't have any other impact apart from reducing memory requirements during sync.

PiperOrigin-RevId: 441789775

GitOrigin-RevId: 78adf7564c3b28173e3b5e0aa595f3c8b6a0fc39