[2022.1 compat] Make `BlazeGoImportResolver` first in the `GoImportResolver` extension points
In 2022.1, `VgoImportResolver` returns an empty list instead of null when queried about a path. `GoPackageUtil` goes through the `GoImportResolver`s and returns the first non-null value a resolver provides. That's why this change made `GoPackageUtil` return this empty list instead of proceeding to the next resolvers and returning the value provided by `BlazeGoImportResolver` as before.

This CL updates the order of `BlazeGoImportResolver` to be first to ensure that its value will be returned.

PiperOrigin-RevId: 443095722

GitOrigin-RevId: 493f9436daaf115d2bbb89e75d425c1ca63799e9