Validate sync when only a workspace module exists
AndroidModels are added to facets in BlazeAndroidProjectSyncer#updateAndroidFacetWithSourceAndModel, which is called during the final phase of the sync process. When a project is first opened, the initial directory update phases tries to validate that step of the sync, but at that time only has a single module (.workspace) without an AndroidModel, which results in a misleading error surfaced in the IDE, although the sync will still complete successfully, and subsequent directory updates pass validation.

This change accepts a sync state as valid in the case where there is on a single Android facet, which is associated with the ".workspace" module.

PiperOrigin-RevId: 451172080

GitOrigin-RevId: 3ae7f6af96f2c0e04c40e7bef8dafb72cffc1659