Expand task tree when non-leaf top-level task is added
In the task tree view of the Blaze tool window, the expected behavior is for the tree to auto-expand up to one level below top-level tasks. The previous logic only expanded if a task was added to a top-level task already in the tree view, and did not account for the cases when the async update is such that a top level task is added along with children, (and the TreeModelListener insertion handler does not fire for new nodes added to collapsed paths). This led to cases where the tree did not automatically expand and remained collapsed at the top level.

This CL adds logic to expand the tree view when a non-leaf top-level task is added.

PiperOrigin-RevId: 451413836

GitOrigin-RevId: bc0121147e4e2328b0af6f7af9296d1a44185231