Blaze window - check if removed task is selected task
NPEs could occur if a user's IDE was configured to not show the Blaze console on successful syncs, and enough syncs were run such that what had been a selected task was removed from tree model while the window remained closed. At this point, if the Blaze window is opened, the bind method of the TreeConsoleView tries to display a console that has been removed.

This change updates the tree model if the removed task is or is an ancestor of the selected task, in which case the selected task is set to null.

PiperOrigin-RevId: 454018404

GitOrigin-RevId: a99d53205bb2f47f8e973ee493a216a6adbe613b