Add experiment for legacy build sharding of smaller projects
Increases the minimum targets per remote shard to 500. Previously, the effective minimum was 100 (1000 targets being the minimum size for a remote build split across 10 concurrent shards). The value is raised after investigation showed that the overall size of the BEP increases as builds are split into more shards. User-specified shard sizes in blazeproject files continue to take precedence.

Adds a boolean experiment, enabled by default, to retain the sharding strategy of splitting targets across at least 10 build shards. This ensures that the sync experience remains the same for smaller projects.

Adds tests to the computeParallelShardSize method for both cases in which the experiment is enabled or disabled.

PiperOrigin-RevId: 455705132

GitOrigin-RevId: bd85d3de41dd22009aef4c07948fa5ca38b2ae57