[fix] ensure AsyncOutputProcessor deterministically processes stdout (#279)
1. When child process exited, `isRunning` flag was immediately set to
false, and the handlers were not called due to if(isRunning.get())
condition. It was very likely that a bunch of last lines were ignored
due to this.

2. In case stdout handlers execution took too long, the exception has
been silently ignored, despite very short timeout. So it was likely
that a line could be skipped even if the process was behaving correctly.

GitOrigin-RevId: dab8863c321b36b795b135b914f8b5f0a0a4be58