Update new project wizard to support loading from a project proto.
The wizard is updated to add a "Generate from project proto" option which will read from a binary proto file. Protect the new option behind an experiment boolean to ensure it's disabled by default. When the new options is selected, the "Edit Project View" wizard step is still shown but is unnecessary. Fixing this requires larger changes to non-prototype code, so leaving it as is for now seems ok.

Update BlazeImportSettings to include a reference the the proto file to be loaded. When the new option is selected, the existing project view file will be null, which will currently break a bunch of downstream functionality. This will be fixed in follow-up CLs.

The actual importing of the project from the proto is not done yet.

(cherry picked from commit 600b0d54db2e24809b3132ea4b6b8a705f03c0a2)

GitOrigin-RevId: 07a25f4399f65969a7d6569715bf4b3aaecb5e7d