Update `release_oss_bazel_plugin.sh` script
This CL updates the script to be used for releasing the OSS Bazel plugin for IntelliJ, CLion and Android Studio from a given release branch. The script will now also try to build and release the plugin for all the IDE versions supported for the given product if the version argument is not passed.

(cherry picked from commit 5e785017a9275e93784b27cac4a217950a69c34c)

GitOrigin-RevId: 56a369788a46b525676f698ffc295f110af91422