Cleanup how BEP is deleted after parsing test results
BuildEventProtocolTestFinderStrategy is responsible for parsing test
results from a BEP. A surprising side effect of the findTestResults
method in that class is to delete the BEP file at the end.
This is unexpected from a method with this name, but it also doesn't
make much sense because there already exists a cleanup method in
the same class (introduced in a later unknown commit).

The current CL moves the deletion of the file to the cleanup method.
Both findTestResults and deleteTemporaryOutputFiles are only called
from a single location - BlazeXmlToTestEventsConverter, so there
are some minor changes in that class as well to make it slightly
easier to read.

(cherry picked from commit 871a699f69abc908d6f8b5a7ea2b69d736e8da55)

GitOrigin-RevId: e9a5423edc5238fb3534d5e57d660dea3218cfa4