Allow running single Scalatest tests in Intellij. (#3744)
* Allow running single Scalatest tests in Intellij.

ScalaTestContextProvider is responsible for figuring out which parameters
to pass to the Scalatest runner. It contains code capable of deciding which
test class to run, but not which individual test case in the class to run.

This commit replaces the Bazel plugin's own logic for this with a call to
the corresponding logic in the Scala plugin, which is capable of deciding
both the class and test name.

When no test name is detected (e.g. due to running a whole class), the
plugin will continue passing the class name in --test_filter as before.

When a test name is detected, the plugin will instead set --test_arg=-t
--test_arg="testName", and leave --test_filter out. This is because
rules_scala turns --test_filter into the -s Scalatest runner parameter,
which selects a full class for execution. Passing both -s and -t would
cause Scalatest to run both the full class, as well as the selected test.

Fixes #3425

* Address comment

* Specify both class and test name when running a single test.

* Use -s directly instead of passing --test_filter

Make the BlazeFlagsModification set -s always, and -t if appropriate.
Remove the code path that sets --test_filter.

* Add tests for the flag modifier.

* Also escape backslash

* Fix bug where running a single test then prevented you from running the whole test class

* Remove broken handling of backslashes, leaving it for a future issue

* Handle the case where more than one single test case is selected

* Rename a method as requested

Make two fields static