Add build flags from BuildResultHelper to TestUiSession
With https://github.com/bazelbuild/intellij/commit/31d906609873bb4c145214f3589b54d594064018, the responsibility of adding a build flag corresponding to the path of the local bep file was moved from testUiSessionProvider to the buildResultHelper. As a result, while creating a testUiSession, we need to add buildResultHelper.getBuildFlags() to the blaze flags.

The test debugging runConfig uses BlazeJavaRunProfileState and buildResultHelper.getBuildFlags() is missing in it causing the tests to fail when debugged. This CL fixes that.

(cherry picked from commit 42ed55acc88238975aec34210ac264241e421e71)

GitOrigin-RevId: e99a7cfdde757cbc307d3cec123bb31c538d6f16