Add build flags from BuildResultHelper to TestUiSession
With https://github.com/bazelbuild/intellij/commit/31d906609873bb4c145214f3589b54d594064018, the responsibility of adding a build flag corresponding to the path of the local bep file was moved from testUiSessionProvider to the buildResultHelper. As a result, while creating a testUiSession, we need to add buildResultHelper.getBuildFlags() to the blaze flags.

The test running/debugging runConfig in CLion uses BlazeCidrLauncher and buildResultHelper.getBuildFlags() is missing in it causing the IDE to not be able to parse the test results properly. This CL fixes that.

(cherry picked from commit d6cdf7a2f871e9b951a0eb15edd4584a829215b0)

GitOrigin-RevId: 874a9e3f319f92089d2470c360670104c3e77896