Improve error reporting when project view file not found.
Instead of silently returning null, allow the following code to throw a FileNotFoundException. ProjectViewManagerImpl appears to be the only user of this code, so this should be safe. Also, in ProjectViewManagerImpl, don't log a full stack trace as these are prone to being silently omitted by the IntelliJ logging code, instead report the exception message.

This makes it somewhat easier to debug issues when the project view does not exist on disk.

(cherry picked from commit f2fb55baef27ac8bebfd26415ae7f2338b6f15c8)

GitOrigin-RevId: 5f42e597efda5d277b897a4ada5535458c23a1f2