Fix getFileDependencies in the multi-target case, and test that.
Method getFileDependencies is used to determine what needs to be build in order for a given source file to be edited fully. It seems reasonable to return the minumum set of depencencies, to optimize build times and IDE performance.

Prior to this CL, the method would return all transitive dependencies of the source target. This includes all targets that include the source, so returns the superset of all deps needed to build it. Instead, first resolve the source file to a single build target, leveraging the existing behaviour that selects the dep with the fewest dependencies. That find all transitive dependencies of that target, which should be the minimum set of deps needed.

Add an assert to the existing test case for this behaviour. Before this change, this assert would have failed.

(cherry picked from commit fe013a7059e0bdf65af84290f455df7d1b9a79cb)

GitOrigin-RevId: 5eac81a4f02e24a645dd4f44ea8bcba864065925