Support Bazel 6.0.0 (#318)
* Support Bazel 6.0.0

1. Main Repository Target Label Prefixes.
In Bazel 5 and older, the target labels were strigified to "//"-prefixed names. Since
6.0.0, they are stringified to "@//"-prefixed names. Unfortunately, this does not apply
to data coming from BEP. Targets references in BEP events payload are "//"-prefixed.
For this reason we need to add "@" manually after retrieving data from BEP.

2. aspects.bzl: Retrieve runtime toolchain in the new way (Bazel 6.0.0)

3. cpp-project: Old GTest version is incompatible with Bazel 6.0.0, the
new one required cpp version bump (--std=c++17)

4. e2e tests are now adapted to run across multiple Bazel versions (currently 5 and 6)

5. JdkResolver: JDK version was inferred from source version, which is not always true. That's
why it is safer just to pass null.

6. Remove 'version' field from JdkCandidateData

It was based on "source" version of the jdk, which was not always equal
to the JDK version. Now, we always call "java -version" to retrieve the
version number