Make BuildGraph an immutable class.
BuildGraph is now an @AutoValue class with a builder. Two distinct bits of logic have been moved outside of it:
1. Listening and updating functionality, which now lives in BuildGraphHolder
2. Parsing of the blaze query output, now in BlazeQueryParser

The split between BlazeQueryParser and BuildGraph is not 100% clean and is an area of future improvements. Note that BuildGraph is not strictly immutable in this CL, due to the memoization of transitive deps.

This CL paves the way for future work to make incremental updates to the project.

(cherry picked from commit 5f8583daaeff53c0e17cf1002bdb4fcbc1d7c6cb)

GitOrigin-RevId: 2dd4646adac82351655b6c0300d27143c19073b3