Improve error messaging when querysync actions fail.
This is an incremental improvement, but needs a lot more work. In particular, we should be able to distinguish between user-actionable errors, and others, to allow users to self correct problems when that is possible. A lot more work is required to make that possible.

(cherry picked from commit faec20a3e715d792e055f2d0bc1ceb0ee4bd6ded)

GitOrigin-RevId: 9a942e6ac2159acf0c765a94a7a7c3341ccbfcbb