Include additional jars when analyzing a proto rule
The proto toolchain is there as an implicit attribute of
the proto rule, but is not a JavaInfo rule. This rule does
not transitively depend on other JavaInfo rules, so pulling
the transitive jars is contained to whatever the toolchain
adds.

(cherry picked from commit 2b2fc875fc71e7f572134ede8ca9459f3ef781c4)

GitOrigin-RevId: 2c67a5d426de1fe505d9d3e42ad0208e94a21b36