Add support for included bzl files when calculating affected packages.
When editing a bzl file, we need to re-query any build files that include that bzl file.

Add information about includes to the Query.Summary proto, and copy that information from the query output into there. Use that information inside AffectedPackagesCalculator when determining the set of affected build packages. Add tests for the new functionality, and appropriate test data and utils.

Also switch to a Builder class for AffectedPackages since that lends itself nicely to the way in which the class is instantiated.

(cherry picked from commit a5c16e65d4040e1572849aa5f7377bff2b474352)

GitOrigin-RevId: ba60285586047f791c8cc3fda7f1eb872351cfa2