Add support for edits to source files when calculating affected packages.
When a source file is added or deleted, we need to re-query the package that includes that source file. This ensures that when source files are added or deleted, the IDE state is is updated to reflect that.

Also add a PackageSet class which encapsulates the "getParentPackage()" logic previously in QuerySummary so that the logic can be reused inside AffectedPackagesCalculator.

(cherry picked from commit c0f03ce8b90bbb3723ff9828ff412fbc9d06199c)

GitOrigin-RevId: 6e9e5429a88d4a82fffd7ab4461d324c6232ca54