Split BlazeProjectSnapshot into constituent parts with different lifecycles.
Different subsets of the data within BlazeProjectSnapshot are available at different phases of sync, so the structure should suppor that. This CL creates the following classes:
- SyncSpec: the input to the sync process that can be created from the project configuration. This class remained unchanged throughout sync.
- PostQuerySyncData: the state after the query invocation has been made, or after a delta has been applied to that. This class is the input and output to the partial update operation, and also contains the data that will be persisted to disk over an IDE restart.
- BlazeProjectSnapshot: the full project state, created in the last phase of sync from PostQuerySyncData.

These classes are put into the .project subpackage to keep them somewhat separate from the main querysync logic.

The motivation for this CL is to make saving the project state easier, in particular allowing the partial sync code to be reused in this context based on the PostQuerySyncData. That will be done in a follow up CL.

While we're here, also improve some encapsulation:
- SyncSpec becomes a natural place for the logic of determining which project paths can be queried (i.e. contain a BUILD file)
- We also move the "sortExcludesByRootDirectory" logic to the same class.
This moves said code into an easier to test location, though no new tests are added in this CL.

This CL results in the NoopProjectRefresh becoming difficult to maintain in a clean way, so that optimization is removed for now. The main upshot of this is that the graph to project conversion logic is re-run in this case, which doesn't seem like a bad compromise to make for less code complexity. It can be reinstated in future if deemed necessary.

(cherry picked from commit b1aa252bfd8af9ed3ed4d69f250192dca55c2162)

GitOrigin-RevId: 9be8b289fa941650534e7bd06e1eb46e3e59fcbd