[fix] Handle uninitialized BspConnectionService in UI elements' `update` method | #BAZEL-304 Done
[fix] Handle uninitialized BspConnectionService in UI elements' `update` method | # BAZEL-304 Done

Most of the actions done in plugin require BspConnectionService to be correctly initialized. For this reason
we have been assuming that it is initialized always, and we've been considering it as non-recoverable error,
what was implemented in `ValueServiceWhichNeedsToBeInitialized`. As it came out, there are some places where
the BSP's plugin code is executed, but the service might not be initialized. One of these places are UI
Widgets' and Actions' `update` methods, which are called even in a non-BSP project. For this reason, in this
places we need to graciously handle that case.

That's why in this commit we no longer derive `BspConnectionService` from `ValueServiceWhichNeedsToBeInitialized`,
but we keep it's value nullable instead. In places where it causes non-recoverable errors we still obtain the
value with double-bang (!!) operator, but in places where it is possible to handle it, we just do nullability-check.

Merge-request: BAZEL-MR-189
Merged-by: Tomasz Pasternak <Tomasz.Pasternak@jetbrains.com>

GitOrigin-RevId: fc5990db089845cd7432568e95070222690b4e12