Improve the handling of changed .blazeproject files.
When loading a project from the snapshot, check that the ProjectDefinition is the same as what we has last time, and if not throw away the entire snapshot. This ensure that project changes are picked up on IDE restart, which is not ideal (we should do it on sync), but better than what we have now.

Also add internal action to reload the project, so this can be triggered without an IDE restart.

(cherry picked from commit 507beab46540a233db16de5b08840a0bc266c285)

GitOrigin-RevId: b28ce9c1255fc782cdc91df78283138c19d793f5