Add support for task to complete with warnings.
In this case, we show a waring icon ("!" in a yellow triangle) to alert that user that task completed, but there may be problems.

Use this warning state for building deps, when we managed to build some dependencies, but not all, to help ensure that the user is aware of this. Otherwise, the task was shown as successful which is misleading since we expect that some symbols may not resolve.

(cherry picked from commit 4760e7e16e11271d9944b2a6f4bfeedbdb54156b)