Add a version to the query summary proto.
The intent here is that the version would be bumped every time there is a material change to the logic in QuerySummary that would yield a QuerySummary incompatible with one from an earlier version, for the purposes of calculating deltas or downstream project structure calculation.

Set the initial version to 0 such that existing projects are not affected.

Do not perform a partial quary update in the case that the version has changed, falling back to a full query in this instance.

(cherry picked from commit dca32341cb6543e22fe40bcbe35383835cf474fe)