Handle non-Blaze projects and non-workspace files more cleanly.
If the project is not a querysync project, do not disable analysis, and do not show the build dependencies action.

For a file outside of the workspace, do not try to check if it's ready for analysis since that causes a crash.

(cherry picked from commit 95dafcdc50bc0d8301bcc345bc35efd799a8ff6f)

GitOrigin-RevId: b8ba9743e1ed9132750e018150661f449b135d86