Separate OutputArtifactDestination and cache layout concerns
Fetching and caching artifacts is a concern separate to maintaining the cache
directory layout in a way that can be consumed by the IDE.

This change extracts the `OutputArtifactDestination` interface core methods
as a new `OutputArtifactDestination` interface and keeps the original set of
methods under `OutputArtifactDestinationAndLayout` name. This is a first step
towards complete separation of artifact caching from artifact layout.

The interfaces are combined via `extend` to limit the scope of this change.

(cherry picked from commit faf4a502bf5e5239da19ef0319cef5256f16472b)

GitOrigin-RevId: e8ee915d3b70e13d10e4944a6bf5d5142cf8ce10