Don't choose non-existant files to represent packages.
Doing so results in a FileNotFoundException later on.

To enable this, encapulate the workspace root as a Predicate<Path> inside GraphToProjectConverter to allow it to check for the existence of a workspace root relative path. This approach simplifies the testing. Update test code accordingly.

(cherry picked from commit 977f710c5b6e7bedc5d69f298d797ce65492000e)