Add support for optimal partial sync based on which files changed between snapshots.
Introduce a new API in `BlazeVcsHandler` to determine which files actually changed in the VCS between two points in time, represented by a `VcsState` object. Use this information in the querysync library to avoid re-querying packages that didn't actually change. This means that when a `BUILD` file is present in the working set, we will only re-query it's package if the file has actually been modified since the last sync. The same goes for `.bzl` files, where the impact can greater if they are included in many `BUILD` files.

Also refactor some code in `ProjectLoader` to avoid creating a `BlazeVcsHandler` more than once not that it is needed in more than one place.

(cherry picked from commit f4e58e902c680a1909a05ab629d8a5291519398b)

GitOrigin-RevId: d6539878ca752ec6cf1524f64ca3bc65ee03b996