Improve error reporting in query sync.
In QuerySyncManager, add appropiate issue parsers to the tool window scope, to extract error from the build output and show then in the "Blaze Problems" view.

This alone surfaces problems with the use of effective workspace roots, whereby the file referenced in the blaze output is not in the users workspace. Fix this by replacing the effective workspace root that the command was run in with the real workspace root in the build output. This ensures the clicking the problem shown in the problems view opens the correct file.

Modify the `BuildResultHelper` to allow this new logic to live in the appropiate layer.

(cherry picked from commit df0fcbcdb722cf411c4504d5606ab4ea7ad57597)

GitOrigin-RevId: 18638b8479d0fd26599c5e283909b8d2b6e40a8a