Add "Build dependecies for open files" action.
This build dependencies for all project files that are currently open in editor tabs.

Refactor some existing code from BuildDependenciesAction so that it can be shared with the new action. Update some other codepaths to allow a list of paths to be passed instead of a single one.

The new action is triggered by keyboard shortcut ctrl+alt+shift+D, which is the same a the existing build dependencies action but with shift too. It's not yet exposed via any other UI.

(cherry picked from commit f678d7d01e3df61caa1a81c424711a30567190a1)