Prevents opening projects with different syncs
Instead of failing at sync time, the IDE will
present a pop-up offering to convert. The
conversion is always set to fail and is only
there to prevent the project from opening.

There is no other mechanism to prevent projects
from being opened.

While we could do a conversion, we don't want
to do that until we are sure users won't need
to go back. The while point of the conversion
would be to delete all downloaded libraries,
but doing that would force users that want
to go back to the old sync to perform full-on
legacy sync.

(cherry picked from commit 4931f0363d2f16c4431df10a61422db407c4141e)

GitOrigin-RevId: 8ee484e86557289d7094d856aec836c0fbf5aa9b