Fixes query sync project decorator
... to have a space between the name and the dependency count,
and to use the original text, instead of the file name.

If the file was a java class, the IDE would have removed the .java
extension, while we were just using the file name. That caused not
just the number of dependencies to be gone, but the name itself
would change. This change fixes that.

(cherry picked from commit 5947c49f6f54ca05fe6cd1e863b268be4928a68b)