[performance] Don't call `buildTarget/javacOptions` if `workspace/libraries` is available
In the past, we used to use javacOptions endpoint to retrieve classpath, which was then
used to reconstruct dependency structure of external libraries (each classpath entry has
been mapped to a single library)

Once we started using an experimental BSP-extension `workspace/libraries` we no longer need
classpath, hence we don't need javac options.

Apart from that it turns out that the JavacOptionsItem data structure significantly impacts
memory footprint of IntelliJ. For a 20k-targets project, the retained size of these data
structures is 439 MB, out of total 1515 MB.

The 20k-targets project mentioned above has been generated with this script.
https://github.com/JetBrains/bazel-bsp/blob/3360353d9a2d12931972ea3f6bf80e2b04cfa237/bspcli/BUILD#L30

For that reason, this change causes intellij-bsp to skip `buildTarget/javacOptions` in
case the `workspace/libraries` endpoint is available


Merge-request: BAZEL-MR-363
Merged-by: Tomasz Pasternak <Tomasz.Pasternak@jetbrains.com>

GitOrigin-RevId: 09f23f802ee8d9e56058c428eca69bffc2aa51e7