Initial prototype of the new rules for ASWB
This is based on the setup we have on repo.
It removes the need to use DevKit for plugin development,
and uses a plain rule to run and debug the IDE. This implies
we don't need to set up a plugin SDK and it will use what is checked in.

(cherry picked from commit 98dc63ae60b8e9315a7d69b67fbb1552c99c4940)

GitOrigin-RevId: 78165e7dc2ae516d8b3ab6ff71819c3e3bf14879