Improve blaze console logging for user-canceled builds
There are two cases considered when build is canceled during the building dependencies phase of query sync.

1. The build is executing as an external task, in which case the command exits with a -1, resulting in a thrown BuildException.
2. A remote build is requested, in which case cancelation results in an interruption exception when parsing the build event stream.

This change adds logic to the blaze context error handling to check for cases of user-canceled builds. In such cases, a message is logged to the console and the tree node correctly shows the build as canceled, instead of the error indicator that would show up previously.

(cherry picked from commit 5ebe2bf3ab69ad456ffa691fd28946b90624721d)