Rework mapping from source -> target.
Instead of keeping a map of (source target) to (rule that builds it), keep the opposite map: i.e. a multimap of (build rule) to (all source targets). In BuildGraphData, also keep the inverse mapping as a memoized field.

Move that logic for selecting a single build rule for a given source such that it's done on demand, rather than statically at graph construction time. This allows this CL to preserve the existing behaviour.

This unlocks more features, such as allowing the user to choose which target to build for a given source file if it's included in several.

(cherry picked from commit 59afdfe8012f11b38866da47493a9d338b8cf3bd)

GitOrigin-RevId: 80ac89f020d6b0861c42194a257ddbbd5f9573f0