Make the default projectview empty during autoimport and add some comments there (#5162)
The difference between autoimport and regular import is that, during autoimport the user
has no chance to edit the project view file before the import actually happens. Hence,
if a repository is large, the current default "directory" value (".") caused the whole
repository to be traversed.

This means the user sometimes has to wait a long time only to reduce the import scope.
It would be better to import nothing by default and let the users choose what they want to
import.

GitOrigin-RevId: 1edcb4fb828253b3ddae501c788a76407a2e136b