Delete `RetainedHeapLimiter`.
It has been superseded by `GcThrashingDetector`, which is simpler and has proven to be more effective.

Rename the flags without the `experimental_` prefix, keeping the old names to aid in migration. A default value for `--gc_thrashing_limits` is applied in order to keep the existing behavior that specifying a value < 100 for the threshold activates detection.

(cherry picked from commit f28cce1f847d5ea01f77181023ec5b46cddbda98)