Fix build exit code handling.
The ExternalTask code previously called context.setHasError() whenever the process exits with a non-zero exit code. This leads to unexplained error states being reported in many contexts. In the case of build invocations, we have explcit handing of the exit code at a higher level in the stack, and calling setHasError() in this case if usually wrong.

Add an option to disable the default behaviour that can be used in such cases. The existing behaviour is left as-is by default since there are a large number of other codepaths that may rely on that.

(cherry picked from commit a0185844b58e156204764acc630dd9521e3c0154)