Add UI to select a single target to build when there are several.
Instead of silently selecting a target to build when the source-to-target resolution is ambiguous, show a list of targets to the user and require them to choose which one to build. Use a popup window for this when the user performs a "build dependencies" action. If there is only one target (the common case, hopefully), this is skipped to the behaviour is as it was before.

Start using `BuildDependenciesHelper` in `QuerySyncInspectionWidgetActionProvider` to ensure new and existing logic can be shared.

There is some subtlety inside `BuildDependenciesForOpenFilesAction` to try and keep it working as well as we can while keeping the UX clean:

- If there are multiple open files which resolce to a single set of targets, use the chooser as per a single file. This ensures that the case of multiple open sources that are part of the same logical target work as expected
- If there are open BUILD files (that resolve to multiple targets, where we want to build all of them), use those to disambiguate any source files
- If there are multiple source files that resolve to different ambiguous targets sets, ignore them for now and emit a warning in the log.

(cherry picked from commit 5e6295edb7ccebf1168bac75b6619965417c58a3)

GitOrigin-RevId: 28f9004f165e767c26dbf060f3db6e155b4a7429