Update Beta channel from 2022.3 (G) to 2023.1 (H)
For the intellij.blueprint changes, removes --config=java_runtime_sts".
The flag is not used on the current Linux canary which is 2023.1. I think
that means it is not needed, and therefore it is removed for beta with
this change that brings 2023.1 to beta. I also think if it's not needed
for Linux, it's not needed for Mac, so the unnecessary one is removed
from mac canary as well. If wrong, presubmit failures will tell us.

(cherry picked from commit dafff3ceadef4d7d4cc57081958293fa12ec758f)

GitOrigin-RevId: 2f52434ca8ee0c1154a8cadfc48c3b8a1624efe3