Add proto toolchain dependencies to sync logic
Adds proto toolchain attribute to the QuerySummary proto construction and bumps the proto version.

For sync to add these deps to the proto targets, the rules are added to the set of java-specific rules, which is now not mutually exclusive with the always build rule types.

When checking for in-project deps, use the query proto summary, which ensure that the proto_library dependencies are not added.

(cherry picked from commit 9fa2e672609bb11ce7e08b5d5528e7ef6fea4ff7)

GitOrigin-RevId: ebe8abb077aefb5651ac88ed5616addabb8255cf