Introduce `RunCommand` flag `--run`.
If true (the default), the `bazel` client proceeds with a `RunCommand` as normal, running the built target using the command line specified in the `ExecRequest` returned by the `bazel` server.

If false, the `bazel` client skips running the built target. This makes it possible for a process invoking `bazel` to call `bazel run --norun` and subsequently use the `ExecRequestConstructed` event on the BEP to run the built target itself.

Note that this does not apply to `--script_path` runs since the bazel server does not return `ExecRequest`s for this flag.

(cherry picked from commit 4c712a51dbb7856ed9f935dcb48ed8aca8d1f6c6)

GitOrigin-RevId: 9be5a82c55823b5474cb319ff1cf5152bc880f9d