Prepare for 2023.3: Move some tasks to background thread. (#5446)
Since the commit below, FSRecords update operations are forbidden in EDT thread.
https://github.com/JetBrains/intellij-community/commit/78f74745498f981d66733bcba315f85cc86c11c6

For this reason, we delegate a few method calls to Background Thread in order to make
sure it will not crash. We do it via a synchrounous BGT call, so
EDT still stops and waits until the aciton is completed, like it was before this change.

GitOrigin-RevId: 1830abd37b78e2614a5fc673e181ff6c5c54c23b