Add an entry point for registering `BlazeProjectListener`s.
To support C++, a `BlazeProjectListener` will need to be added inside the `plugin/cpp` library, which we can't depend on from `plugin/base`. To allow this code to register as a project listener, add an entry point for this.

Also convert the existing listeners to use this entry point for consistency. Since this changes the way that it is constructed, add appropiate getters to `QuerySyncProject` to enable it to acquire all necessary dependencies.

(cherry picked from commit d1db75988c035ff4c0ff98ce1400ec659faac017)

GitOrigin-RevId: 8c7fc7680cf0828e00608b57e64209dc9e329399