Add support for building deps of cc targets, and reading the build artifacts.
In this CL:
- When building deps for a cc target, request the appropriate output groups from the aspect
- Copy the generated headers into a new direcory in the cache, ensuring that their paths and filenames are preserved since they are referenced by their path in `#include` statements. Add a new `CacheLayout` to do this, `ArtifactPathCacheLayout`.
- Add a proto definition corresponding to the `.cc-info.txt` files produced by the aspect
- Add code to download and parse the `.cc-info.txt` files using this proto definition.

This means that when building deps for a cc target, the build will be run and artifacts downloaded, but as yet we do not do anything with those artifacts. This will come in future CLs.

(cherry picked from commit c61969cb0f9ce46d9528f2404e13263de5164ffc)

GitOrigin-RevId: 29260591f6160a7f250bfde5d5deb96ad0415f3b