Make //clwb(plugin) and //cpp(plugin) dependency on cidr.lang optional: plugin xmls split (#5473)
# Discussion thread for this change

Issue number: #5472

# Description of this change

Long story short: `com.intellij.modules.cidr.lang` is going to become optional in CLion at some point. All the usages of code exported from `com.intellij.modules.cidr.lang` are moved to optional plugin xmls, so plugin won't emit lots of exceptions and break the IDE if the mentioned module is not available. 

A little bit more detailed here. The CLion team are working on deprecation of the old language engine. Parts of it which are not going to be deprecated are moved to `cidr-base-plugin` library, which is a part of `CIDR Base (com.intellij.cidr.base)` essential CLion plugin, so any symbol from `cidr-base-plugin` will always be available. Though that does not apply to `c-plugin` library which is a part of `com.intellij.modules.cidr.lang` module. 

At this point the goal of this change is to make Bazel for CLion tolerant to disabled `com.intellij.modules.cidr.lang`. We're going to contribute further to adjust existing code which is moved to `*.oclang(.*)?` packages in the review to preserve the features when only the new language engine is available.

# Individual changes

* Replace cidr.lang code with cidr.lang.base code
* Extract cidr.lang dependent code in //cpp to optional runtime dep
* Extract cidr.lang dependent code in //clwb to optional runtime dep
* Extract cidr.lang dependent code in //clwb to optional runtime dep