Add placeholder code for populating CC information in the project proto.
Add a new proto message to `project.proto` to hold the cc workspace information (currently empty, follow-up CLs will populate it).

Create a `CcProjectProtoTransform` and corresponding `CcWorkspaceBuilder` which will update the project proto with this new proto message based on the build graph and the output from a cc dependencies build held by the artifact tracker. Add an appropriate getter to `ArtifactTracker` to allow this data to be accessed. Again, the logic in here is currently omitted and will be added later.

Add a new `CcDependenciesInfo` class to hold the output from the aspect, to reflect changes to the output from the aspect made in precursor CLs. Add an appropriate getter here, and other utility code elsewhere to be used by `CcWorkspaceBuilder`.

(cherry picked from commit 19b2ebe3e852cee6e0dc9e91fcf8dfecba90ed47)

GitOrigin-RevId: a59369f0084c1655db8be6007db43ec0a91469a3