feat: Support gutter icons in `java_test_suite` (#5579)
java_test_suite is a common way to define java tests for users of https://github.com/bazel-contrib/rules_jvm. Running them is roughly trivial as it acts just like a test_suite that generates java_test targets. Just making the plugin aware of the target kind is enough to make it work. All of the individual java_test are indexed by the aspect normally.

---------

Co-authored-by: Thom Jones-Low <tjoneslow@apple.com>

GitOrigin-RevId: 02657dd4eafb03539e14efc5e4dab32ffeabe130