Ensure NDK support is enabled after building deps for cc targets.
The legacy `BlazeNdkSupportEnabler` will disable NDK support if C is not in the set of active project languages. Whlie it's not clear if this functionality is still needed, it seems safer to keep it.

To ensure it works as expected:

- Modify query sync code to augment the list of active languages based on what's in the project. For C/C++, this means addng C to the list of active languages if we have any CC targets configured. Add the set of active languages to the project proto to enable this, and populate it appropriately.

- Make `BlazeNdkSupportEnabler` query sync aware, specifically make it implement `BlazeProjectListener`. This is necessary since the set of active languages now changes after a build deps action, not just after a sync.

(cherry picked from commit bc8971ebe60a89752772db26f115f8d4268cafb6)

GitOrigin-RevId: ca8c42ee663af5de5dc9759c7cb991d2f362be7c