Pass -include compilation flags down to OC language model (#5645)
For some reason particularly -include was marked as unsupported back in 2018,
and we don't know the reason due to the lack of knowledge. To address any
possible issues with the change behavior a new registry flag
bazel.cpp.sync.workspace.filter.out.incompatible.flags is introduced. It is
false by default, so we enable -include support for everyone and the old
behavior can be brought back by enabling the registry flag.

GitOrigin-RevId: 526d5afccc0e000e032846dd230280cb136a0228