Add an option to bypass source-transformation build rules.
Build rules than perform some form of tranformation to source files before building them generally do not mix will with the IDE, since the IDE does not know about the transformation, and the original sources are not recognised as sources because they are not compiled directly.

In some cases, the transformation is not required to get the IDE to successfully index the source files. Add support for doing this on a per-rule basis, by adding a new tag `ij-ignore-source-transform`.

When applied to a source transformation type build rule, it causes two things to happen:
- When determining the sources for a target, we treat the tagged target like a `filegroup` - i.e. we expand it's sources in place, treating them as if they were listed directly in the dependant targets `srcs`.
- When building depdenencies, we ignore generated sources that come from a tagged target. We do this to avoid the sources being included twice by the IDE - the originals, and the generated transformed sources.

For example:

```
java_library(
  name = "consumer",
  srcs = [":processed_sources"],
)

process_sources(
  name = "processed_sources",
  srcs = glob(["*.java"]),
  tags = "ij-expand-srcs",
)
```

This would result in the IDE treating the sources for `processed_sources` as the sources for `consumer` itself.

Whether or not this is the right thing to do depends on the nature of the transformation itself. If the transform is required in order to yield valid java source, it will not work. If the source can be compiled without being transformed, using the new tag may be appropriate.

In order to implement the required logic inside the apsect, we now must also traverse `srcs` edges so that we can determine if generated sources come from a target that specifies the new tag.

(cherry picked from commit 9ee4b37a16684d1f121ec9dfa5a5eb572b58e5f5)

GitOrigin-RevId: 061a378b760aeec49e22433fa79b21923801ae1a