Handle errors in `.blazeproject` files more gracefully.
Update `ProjectViewManager.reloadProjectView` to throw rather than returning null if it fails to parse the project view file. Update test implementations to now that this method is not `@Nullable`.

Update the query sync load/reload/sync flow to handle the fact that checking or changes to the project view can now throw. The exception handling is now more centralised, and more logic is moved inside the various `run(...)` calls as a result.

Now, any errors in the project view are reported in the blaze problems view, and loading/syncing will terminate if we fail to reload the project view.

(cherry picked from commit aa454a33e29301e6639a182ff10f195d41204e0a)

GitOrigin-RevId: d6971645c32bb0453eb568d1a21c4e1312fb48c6