Add special case support for single line generated java files to the package statement parser.
Since the parser is also used for generated source files in some cases, we can't assume that the input is a "normal" source file with a package statement on a line by itself. Generated source files sometimes have the entire content on a single line, so add support for that.

We do it this way to avoid using a full java parser which add extra complexity (though maybe not too much?).

(cherry picked from commit 426ba269114cbe914507cc8b4686e683bdfc105f)

GitOrigin-RevId: 8ce401345fdc880dff41df03fec2bb897f500645