Set test sources bit correctly for generated source jars.
Generated source jars should have their test bit set according to the target that generated them. Failure to do so results in the IDE failing to resolve references from the generated code to regular code in the same target, since the IDE thinks the generated code is not test code (so cannot access test code).

This requires changing the way that we handle generated source jars. Previously, we wrote them all to a cache directory, then used the directory contents to update the project proto. Doing it this way means we loose information about the origin of the jar file, so we cannot map it back to the target that generated it to determine if it's a test target.

Instead, take the set of generated source jars directly from the `ArtifactInfo` protos that we have from previous dependency builds, and then map the build artifact paths to cache paths using the cache layout. This allows us to look up the "is test" but using the corresponding build graph data and the project definition therein.

(cherry picked from commit e64bf063af07dccd1f7d9a8194b09cbbbc32ccb7)