Exclude `no-ide` targets from `ProjectSourceToTargetFinder`.
This is an old class built for legacy sync, which therefore assumes that `no-ide` targets are excluded from the list of targets. Exclude them explicitly for query sync to ensure that it behaves as expected, and does not change its behaviour when enabling query sync.

Also add the target tags to `ProjectTarget` to enable this, and a simple test case for this.

(cherry picked from commit 852f8123835af9ee47f409412fd9aff76098d7be)

GitOrigin-RevId: c93fbfdee4a0651a2a4e67b24e280b7347023246