Move the definition of the protos output by the aspect for java into the querysync library.
Also move the corresponding wrapper class, and rename it to make it clear that it's java specific.

This will allow the querysync library to depend these things, in turn allowing more of the build deps logic to be in the library.

(cherry picked from commit 3e35774cbbc7a5164e2f00e51f319be67c4dd56a)

GitOrigin-RevId: 855e65c381a2cabaaab41800c32c70ec404e6e39