Include the blaze-out relative path in local file ouput artifact equals implementation.
These objects are used as map keys. It is also possible for the same local file artifact (being, in this instance, a checked-in source file) to appear twice in the BES, if it is used by multiple configurations. Using just the file path for the `equals`/`hashcode` implementation results in duplicate key errors in that instance. Including the blaze-out relative path also fixes that, and also makes the implementation more consistent with `RemoteOutputArtifact` which implements the same interface and is used in the same way.

(cherry picked from commit 372339b9599157997dbebf201f5b1907189e72ed)