Update intellij plugin to use modern provider
Also update the tests to reflect the inclusion of providers. The test function `hasAndroidIdeInfo` was not actually checking for `AndroidIdeInfo` but instead for the legacy `android` provider, so now it's updated to check for the correct provider.

(cherry picked from commit eeb20cd4dd7c7e1fd28c5a6bd9109130f5eeae43)

GitOrigin-RevId: ca0896db12762c4600b4c5466a99d7657b0f1cc8