Split `DefaultCacheLayout` into an unzipping part, and the fallback behaviour into a key based layout.
This brings `DefaultCacheLayout` in line with the newer layouts, where each one has a single purpose and we combine them with a `DelagatingCacheLayout`.

This is a pure refactoring CL, all existing behavior is kept even if it's not clear if it still makes perfect sense.

(cherry picked from commit 7fff27ad1c033f81a0e1fb635854caa0d460ddbc)