Add a `ModificationTracker` for query sync projects.
The tracker is notified whenever the project structure changes. Use this as a dependency when creating cached results from `QuerySyncNavigationPolicy` so that cached values are refreshed whenever the project is updated. This ensures that stale values are never returned (e.g. if a value is cached before the project is loaded, or before dependencies have been built for a target).

(cherry picked from commit d33f65e3bd01a09986393f113236d179912a8de4)