Fix warnings when syncing a workspace containing a bazelignore file (#5873)
* Fix warnings when syncing a workspace containing a bazelignore file

The warnings look like this

WARNING: Pattern '//dist/...:all' was filtered out
by ignored directory 'dist'

They occur because the plugin passes paths from bazelignore into
"bazel query" as exclusions. Bazel warns about this, as excluding a path
that Bazel already knows should be ignored is likely a user error.

The reason bazelignore paths are ending up in the query is that we
want the IDE to mark the paths as excluded from the project. For that reason,
the plugin parses the bazelignore file, and adds the paths to an exclusion
list. That list happens to also be passed to "bazel query".

This commit keeps bazelignore paths in the general exclusion list, but
generates a separate list of paths to use for "bazel query", which only
contains exclusions not listed in bazelignore.

* Add test

GitOrigin-RevId: a01b4228e0d30e0397b4a69470121175bcf866b9