Report unresolved symlinks as such in the execution log.
In addition, check that an output is of the expected type before emitting it, and otherwise deem it nonexistent. Strictly speaking, this is a breaking change for the old (expanded) log formats, but any such spawn will later cause the respective action to fail, so it doesn't materially affect the ability to analyze cache misses.

(cherry picked from commit 59f7e02f1b52224096622e6ef7797eabf17fb907)