[feature][experimental] Basic JPS compilation support + 241 EAP3 update | #BAZEL-695 Done
The JPS compilation feature is hidden behind the feature flag bsp.jps.project.compilation.
Two build actions are introduced to distinguish whether to build with build tool or JPS.
Also, this PR will also be in charge of updating the platform to 241 EAP3 due to a lot of breaking changes + complications with workspace model usages.

Merge-request: BAZEL-MR-751
Merged-by: Xuan Son Trinh <xuanson.trinh@jetbrains.com>