[maintenance] Refactor some spaghetti and kotlinize some files in the init flow
First part of my refactoring aiming to simplify the initialization flow. The most important thing is removal of many lateinit fields and init methods. Also it's possible that trace logging will now be enabled, as previously the result of the call to a builder method which enabled it was ignored. Also errors will now be properly logged with a stack trace.

Merge-request: BAZEL-MR-801
Merged-by: Andrzej Gluszak <Andrzej.Gluszak@jetbrains.com>

GitOrigin-RevId: 30dfadcef1e74e91bd4c90054d34310a80f96b32