Add `BuildArtifactCache`.
This is a simple store of build artifacts keyed by their digest. We add to the cache on demand and provide access to artifacts that are already present, or in the process of being downloaded. We also maintain the last access time for each artifact (using the filesystem attribure) which will, in future, allow old artifacts to be removed from the cache.

It will be used as part of the upcoming new artifact management logic but is as yet unused.

Add some tests for the core functionality. Also move some test utility code into the shared package (from querysync) to allow it to be used in these new tests.

(cherry picked from commit 91a862c382ca9cb9084ccf36dc8e5a47b7a761d4)

GitOrigin-RevId: 640315b69494f287b49ffc949585f87a162825a9