Make `ProtoStringInterner` static.
Requiring an instance to be created is not necessary and makes using it a little more awkward in some cases. We never expect to use it with an interner other than that provided by `StringInterner` anyway.

(cherry picked from commit c64c19a9454bb439f0c485bb848d370fb42c8937)

GitOrigin-RevId: 4cd68f57b3eb6c74109dcc3c5378cea26d34fa6c