Intern protos strings and derived paths stored in artifact tracker state.
Intern the proto strings when reading protos after a dependencies build, and again when restoring the same data from disk. In `JavaArtifactInfo`, also intern the paths we derive from the proto contents. It seems to be worth interning the paths themselves, since they memoize the `toString()` representation internally which we can't memoize directly.

Based on a little experimentation, this reduces the size of the artifact tracker state to about 5-10% of what it was before.

We add a path interner to the `StringInterner` class, and also rename it to `Interners` since it now deals with more than just strings.

(cherry picked from commit 41d78a51287c783734f0715c56173ff14f5d9993)

GitOrigin-RevId: 06579025bd1747a89d2e2e86e56ffe39a01397f0