Rollforward of https://github.com/bazelbuild/intellij/commit/eeb20cd4dd7c7e1fd28c5a6bd9109130f5eeae43: Update intellij plugin to use modern provider
Also update the tests to reflect the inclusion of providers. The test function `hasAndroidIdeInfo` was not actually checking for `AndroidIdeInfo` but instead for the legacy `android` provider, so now it's updated to check for the correct provider.

NEW:
- removed uses of --experimental_google_legacy_api; this was made possible by unknown commit, which exposed AndroidIdeInfo without the flag.
- backward compatibility for old style providers

(cherry picked from commit a97cc5e023b5e958400529e10c3e9c65dea7e271)

GitOrigin-RevId: 6573a71092850069a712ad60d374cb91eb27d353