Allow non-empty package prefixes when appropriate.
Update `SrcJarInnerPathFinder` to specify whether or not a non-empty package prefix is allowed for each call to `findInnerJarPaths` rather than once for the instance. Update call sites accordingly based on whether or not each case can support a non-empty prefix. When srcjars are added as source tpo the project we do support this, whereas when they are added to a library we do not. This reflects the support in the IJ API.

(cherry picked from commit 53d1c7c31985514416fcb28f07310d45cecd4b53)

GitOrigin-RevId: 60d25d92bd57ab9bc522ce6729d47877a0d3b3e2