[maintenance] Create more detailed metrics for applyOnWorkspaceModel
The existing metric `apply-changes-on-workspace-model` that measures the `applyOnWorkspaceModel` function may be not representative in the details of what kind of wait do we have. This metric can be split in three:
1) Apply `replaceBySource`
2) Waiting for the write action
3) Apply workspace model
With this change we'll add time tracking for 1 and 3. Write action await can be calculated, but not explicitly measured.
Also, it's important to understand that 3 happens under write lock, thus causes freezez

This change is a part of BAZEL-317

GitOrigin-RevId: 58b5e3e2868a912400e3df3bd45f727907c56783