[feature] Go support | #BAZEL-749 Done
Add basic Go support in BSP plugin under a feature flag.
This PR is mostly a rewritten version of my team's bachelor thesis project 2023/24.
Previous PRs to archieved repositories
- bazel-bsp: https://github.com/JetBrains/bazel-bsp/pull/507
- intellij-bsp: https://github.com/JetBrains/intellij-bsp/pull/43
This is not a full Go support, there are still some issues to be fixes and features to be added.
Features implemented:
- automatically set GOROOT
- turn on Go support in modules
- recognise Bazel internal and external imports