[maintenance] single place for (almost) all testing rules
I moved junit5 (vanilla) rules from server to a top-level package, so it can be used in all the subprojects. Also, I moved the junit4 rules used in tests with platform fixtures there, so they can be used in other places if needed. In the followup PR I'll do the same for existing intellij junit5 rules and get rid of the suites to improve caching

GitOrigin-RevId: 1b9f1a2317c6093d307f8b4743d96ec91d68fd9f