Initial update to lexer implementation
This commit is the first step to create a new implementation of the Project View lexer. Previously it was implemented using JFlex, but the new lexer isn't going to use it anymore. This implementation needs ProjectViewKeywords to properly parse the LINEAR_KEYWORD and SCALAR_KEYWORD tokens, but it has yet to be imported. The ProjectViewLexerBase is an implementation of the parsing process. The next commit is going to add an extension of the LexerBase abstract class called ProjectViewLexer. It's going to use the new ProjectViewLexerBase to tokenize a char sequence. The Project View token types have also been updated and their number has been reduced.