[maintenance] Update python_info aspect to accommodate Python targets that only have rules_python providers (#210)
As the Python rules are gradually phasing out use of the builtin PyInfo
provider in favor of the ones from rules_python, there could be cases
where a repository is on a Bazel version lower than 8, but the built in
PyInfo is still missing from a given target.

In our repo's case, we are in the process of migrating to rules_py,
which uses the rules_python providers only
(https://github.com/aspect-build/rules_py/pull/311), and not the built
in ones.

This change adds support for rules_python providers in the aspect for
Bazel versions less than 8:
- Bazel <8 - this adds an extra check rules_python in the repo's
external rules. The load statement will only run if present, and when
this is the case, the rules_python providers will be considered as a
secondary option to the builtin ones.
- Bazel 8 - rules_python providers only [existing behavior]