[BAZEL] [fix] Fix flaky `BazelLoadFilepathCompletionTest`
BazelLoadFilepathCompletionTest has proven to be flaky on the buildserver. What that is I am not certain, however I have discovered a possible issue.
The BazelBzlFileService looks for bzl files after each sync and stores the result, this happens on the pool thread. It also used to be run on the pool thread in the init function, combine with instantiating the class on startup it was meant to make sure the completions are available before the first sync from the user.
In tests however, the service was instantiatied, which invoked updateCache on a thread pool and immediatly after the cache was update on the test thread with forceCacheUpdate.
I believe this MR might fix that issue by:
1. Using mutex to atomically check if update is already in progress
2. Don't update the cache from the init so that it doesn't happen in tests

Co-authored-by: Xuan Son Trinh <xuanson.trinh@jetbrains.com>

Merge-request: IJ-MR-169061
Merged-by: Stanisław Solarewicz <stanislaw.solarewicz@jetbrains.com>
(cherry picked from commit 6fc698fe06ac22ea6e5baf1aebeb3e236f2ef5e7)

GitOrigin-RevId: 44ceccc7eea2c8941160528bf7b7bbbfc9fef12d