Disable Target Util Save (#11)
TargetUtils.save writes a large .gz configuration file to file system storage every 5 minutes. This operation is expensive, and we don't use the file anywhere as we trigger sync on startup due to the workspace loading issue. Disabling the save to avoid the IDE freeze when this is triggered.

Add TODO

Update version