BAZEL-2633 new bazel project: .gitignore bazel convenience symlinks
When building a newly-created project by bazel the first time, bazel
creates a MODULE.bazel.lock file, as well as the convenience symlinks.
The lock file usually should be committed after verifying that the
pulled-in versions are suitable, but the convenience symlinks should
be ignored. Therefore add the latter to the .gitignore, at least
to the extend possible, i.e., without taking the exec-root symlink
into account. The name of that symlink contains the workspace name,
which, in the world of bzl modules, is the basename of the checkout
directory (and therefore not constant for the versioned code).

GitOrigin-RevId: 9f8935983bc3a7983dfbe7ed155adf70e733dc13