[BAZEL-1830] Bazel module files: add references for use_extension
In Bazel module files, the global function use_extesion takes the
role of the load statements by returning the value of the referenced
symbol. The arguments are similar to that of a load statement:
the first argument is a label referencing the bzl file to load
the symbol from, the second argument the name of the symbol to be
loaded. Therefore, add references as for load statements.

GitOrigin-RevId: 1cb31b101f04bae7fd108772016d161ad1959f7b