refactoring [text-matching]: extract MatchingMode
Extracted MatchingMode to enable moving the entire NameUtil class from the library code to the platform layer, as MatchingCaseSensitivity is the only thing actually used by matcher.
This move will support parameterizing matchers with a layout-fixing strategy. While this parametrization is a healthy step for a library — shifting state management responsibility to the client — it also increases complexity for IntelliJ Platform clients.

To mitigate this, matcher builders will be moved into the platform core module, which will centrally handle layout-fixing conversions for all IntelliJ clients.

GitOrigin-RevId: 44a3cb761a2052e0d5234cb5b6458b5dde3d6ed9