[BAZEL2677] bazel aspect: do not use --inject_repository
In order to obtain information about the bazel project structure, the
bazel plugin uses a custom aspect. This aspect is made available in
a directory .bazelbsp created by Intellij; it is added to the build
as injected repository with the --inejct_repository option. This
option has the disadvantage that the repository structure differs
from a typical command-line build and hence a switch between a build
in Intellij and one on the command-line invalidates analysis. Avoid
this repository injection (and hence the invalidation of analysis) by
using the fact that the directory lives in the workspace anyway; so
we address the aspect repository as //.bazelbsp/aspects:... as part
of the main repsoitory instead of as @bazelbsp_aspect//aspects:...
with bazel_bsp an injected repository.

GitOrigin-RevId: 8a8ec8afd9c5f0b7b1adc85a4f17814483f4a42e