[BAZEL-2677] bazel aspect: error out if .bazelbsp is .bazelignore'd
After no longer injecting the aspect as additional repository, we rely
on the aspect-drectory being part of the workspace. This usually is
the case, but if the user add that directory to their .bazelignore,
the sync-issued bazel command will not be able to succeed and do so
with an error message not easy to understand by the user. Therefore,
abort the sync with a helpful error from our side.

GitOrigin-RevId: b648cc8165209fcfe2e65d5be3569cddc584fb01
(cherry picked from commit 286b7dfba2d6e83c983d1c78f962b063f91f0b30)