[BAZEL-2677] bazel aspect: error out if .bazelbsp is .bazelignore'd
After no longer injecting the aspect as additional repository, we rely
on the aspect-drectory being part of the workspace. This usually is
the case, but if the user add that directory to their .bazelignore,
the sync-issued bazel command will not be able to succeed and do so
with an error message not easy to understand by the user. Therefore,
abort the sync with a helpful error from our side.

GitOrigin-RevId: eceb2eb8f666f7cea475342b2961e0915933753c