[BAZEL-2677] bazel: improve error reporting on used aspect directory
The bazel plugin, after no longer injecting its aspect as a
separate repository, relies on using //.bazelbsp as a package of
its own. Improve the warning if the user tries to .bazelignore
it by explicitly mentioning that it is the bazel plugin (and not
generically IntelliJ) that needs this directory.

GitOrigin-RevId: 13bef345f62320487b3bc21fc528af47aa68b37b