[BAZEL-2677] bazel: improve error reporting on used aspect directory
The bazel plugin, after no longer injecting its aspect as a
separate repository, relies on using //.bazelbsp as a package of
its own. Improve the warning if the user tries to .bazelignore
it by explicitly mentioning that it is the bazel plugin (and not
generically IntelliJ) that needs this directory.

GitOrigin-RevId: 1b5603558b02b435f8a652a26f0075aabcf3d465
(cherry picked from commit 8a1de9bcd8e6c5af4206bad88e304413ebc1ecc4)