[BAZEL-2737] Change default for `import_depth` to the documented one
The documentation for the project-view paramter "import_depth" (which
can be found in ImportDepthSection.kt) states "The default value
is -1, meaning that all transitive context is included."; however,
when unset the value 1 was used. Fix this by making the implementation
agree with the documentation.

GitOrigin-RevId: ac43a8de741e6aad87c74b8f23d3f5749102da0d
(cherry picked from commit ebd724f29353fbc6713d491925c866dbf5f6eea9)