[BAZEL-2753] Remove left-over WORKSPACE files in the aspect directory
Earlier versions of the bazel plugin used to create a WORKSPACE
file in the generated directory for the bazel aspect. Now
that we no longer inject that directory, a WORKSPACE file is
no longer needed; in fact it can even do harm, das tools like
@rules_bazel_integration_test//tools:update_deleted_packages see
such a WORKSPACE file as indication to add a --deleted_packages
entry to the .bazelrc. Therefore, proactively remove any WORKSPACE
file that might be left over from an earlier version of this plugin
to avoid people adding more --deleted_packages entry that prevent
our aspect package from working.

GitOrigin-RevId: 4e37551957ec48963d5076c8479ee5ea473f5a19