cleanup [bazel]: Use the fact that we obtain absolute labels from aspect
A label in Starlark has the guarantee

  For every Label instance l, the string representation str(l)
  has the property that Label(str(l)) == l, regardless of where
  the Label() call occurs.

and this is asserted for all supported versions of bazel, from
7.0.0 onwards, as well as for the version 6.4.0 and 6.5.0 going
end of life.

In particular, the string representation can never contain aparent
names (as their meaning depends on the repository they are interpreted
within) or be repository-relative. Hence we can always parse the
string representation of a label without knowing the repository
mapping of any particular repository. So drop this unnecessary
dependency in project resolution.

GitOrigin-RevId: fc69891c6cf79ae6ca456be9750239f5cab37cb6