cleanup [bazel] Obtain absolute repository names from aspect
A label in Starlark has the guarantee

  For every Label instance l, the string representation str(l)
  has the property that Label(str(l)) == l, regardless of where
  the Label() call occurs.

In particular, the string representation can never contain aparent
names (as their meaning depends on the repository they are interpreted
within) or be repository-relative. Hence we can always parse the
string representation of a label without knowing the repository
mapping of any particular repository. So drop this unnecessary
dependency in project resolution.

GitOrigin-RevId: a9997e416d7462028bdebaf56415905d8b742975